backend: modularize routes
