+++ backend/main.py
@@ -1,2711 +1,47 @@
-import json
+"""FastAPI application entrypoint."""
+from __future__ import annotations
+
 import os
-import time
-from io import BytesIO
-from typing import Optional, Any
 
-import logging
-from fastapi import FastAPI, UploadFile, Form, File, Header, HTTPException, status
+from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import StreamingResponse, JSONResponse
-from PIL import Image
-# Enable HEIC/HEIF support if available (for iPhone photos)
-try:  # optional dependency
-    from pillow_heif import register_heif_opener  # type: ignore
 
-    register_heif_opener()
-except Exception:
-    # If pillow-heif is not installed, HEIC uploads will fail to decode
-    pass
-from google import genai
-from google.genai import errors as genai_errors
-from google.genai import types
-from .db import (
-    db_session,
-    init_db,
-    Generation,
-    EnvSource,
-    EnvDefaultUser,
-    ModelDefault,
-    ModelSource,
-    ModelDescription,
-    PoseSource,
-    PoseDescription,
-    ProductDescription,
-    Listing,
-    ListingImage,
-)
-from .storage import (
-    upload_image,
-    upload_source_image,
-    upload_model_source_image,
-    upload_pose_source_image,
-    get_object_bytes,
-    delete_objects,
-    generate_presigned_get_url,
-    upload_product_source_image,
-)
-from .tasks import enqueue_pose_descriptions
-from sqlalchemy import select, func, text
-import asyncio
-import uuid
-
-try:
-    from redis import asyncio as redis_asyncio  # type: ignore[attr-defined]
-except Exception:  # pragma: no cover - optional dependency guard
-    redis_asyncio = None  # type: ignore[assignment]
-from .prompts import (
-    classic_detailed,
-    classic_concise,
-    seq_step1_detailed,
-    seq_step1_concise,
-    seq_step2_detailed,
-    seq_step2_concise,
-)
-
-
-def _env_int(name: str, default: int) -> int:
-    raw = os.getenv(name, "")
-    try:
-        return int(raw) if raw.strip() else default
-    except (TypeError, ValueError):
-        return default
-
-
-def _env_float(name: str, default: float) -> float:
-    raw = os.getenv(name, "")
-    try:
-        return float(raw) if raw.strip() else default
-    except (TypeError, ValueError):
-        return default
-
-
-# Config
-MODEL = os.getenv("GENAI_MODEL", "gemini-2.5-flash-image-preview")
-API_KEY = os.getenv("GOOGLE_API_KEY", "")
-GARMENT_TYPE_CLASSIFY = os.getenv("GARMENT_TYPE_CLASSIFY", "1").strip().lower() not in ("0", "false", "no")
-GARMENT_TYPE_TTL_SECONDS = _env_int("GARMENT_TYPE_TTL_SECONDS", 86400)
-REDIS_URL = os.getenv("REDIS_URL", "").strip()
-GARMENT_TYPE_CACHE_VERSION = os.getenv("GARMENT_TYPE_CACHE_VERSION", "v1").strip() or "v1"
-GARMENT_TYPE_CACHE_PREFIX = os.getenv("GARMENT_TYPE_CACHE_PREFIX", "garment_type").strip() or "garment_type"
-GARMENT_TYPE_LOCK_TTL_SECONDS = max(1, _env_int("GARMENT_TYPE_LOCK_TTL_SECONDS", 30))
-GARMENT_TYPE_LOCK_WAIT_SECONDS = max(0.5, _env_float("GARMENT_TYPE_LOCK_WAIT_SECONDS", 5.0))
-REDIS_OP_TIMEOUT_SECONDS = max(0.1, _env_float("REDIS_OP_TIMEOUT_SECONDS", 0.5))
-REDIS_OPERATION_RETRIES = max(0, _env_int("REDIS_OPERATION_RETRIES", 1))
-REDIS_RETRY_BACKOFF_SECONDS = max(5.0, _env_float("REDIS_RETRY_BACKOFF_SECONDS", 60.0))
+from backend.config import CORS_ALLOW_ORIGINS, LOGGER, REDIS_URL
+from backend.core.redis import close_redis_client, get_redis_client, redis_asyncio
+from backend.db import init_db
+from backend.routes import router as api_router
 
 app = FastAPI(title="VintedBoost Backend", version="0.1.0")
-logger = logging.getLogger("uvicorn.error")
-
-_redis_client: Any | None = None
-_redis_retry_at: float | None = None
-_redis_connect_lock = asyncio.Lock()
-_local_singleflight_locks: dict[str, asyncio.Lock] = {}
-_local_singleflight_guard = asyncio.Lock()
-
-
-async def get_redis_client() -> Any | None:
-    """Return a shared Redis client when REDIS_URL is configured."""
-
-    global _redis_client, _redis_retry_at
-    if not REDIS_URL or redis_asyncio is None:
-        return None
-    if _redis_client is not None:
-        return _redis_client
-    loop = asyncio.get_running_loop()
-    now = loop.time()
-    if _redis_retry_at and now < _redis_retry_at:
-        return None
-    async with _redis_connect_lock:
-        if _redis_client is not None:
-            return _redis_client
-        loop = asyncio.get_running_loop()
-        now = loop.time()
-        if _redis_retry_at and now < _redis_retry_at:
-            return None
-        candidate: Any | None = None
-        try:
-            candidate = redis_asyncio.from_url(
-                REDIS_URL,
-                encoding="utf-8",
-                decode_responses=False,
-                socket_timeout=REDIS_OP_TIMEOUT_SECONDS,
-                socket_connect_timeout=REDIS_OP_TIMEOUT_SECONDS,
-                retry_on_timeout=False,
-            )
-            await asyncio.wait_for(candidate.ping(), timeout=REDIS_OP_TIMEOUT_SECONDS)
-        except Exception as exc:  # pragma: no cover - network errors
-            logger.warning("Failed to initialize Redis client: %s", exc)
-            _redis_client = None
-            _redis_retry_at = now + REDIS_RETRY_BACKOFF_SECONDS
-            if candidate is not None:
-                try:
-                    await candidate.close()
-                except Exception:
-                    pass
-            return None
-        _redis_client = candidate
-        _redis_retry_at = None
-        return _redis_client
-
-
-async def _record_redis_failure(exc: Exception) -> None:
-    """Invalidate the active Redis client following an operational error."""
-
-    global _redis_client, _redis_retry_at
-    loop = asyncio.get_running_loop()
-    _redis_retry_at = loop.time() + REDIS_RETRY_BACKOFF_SECONDS
-    client = _redis_client
-    _redis_client = None
-    if client is not None:
-        try:
-            await client.close()
-        except Exception:
-            pass
-    logger.warning("Redis unavailable, falling back to local cache: %s", exc)
-
-# CORS - allow local Next.js dev and same-origin deployments
-# CORS origins from env (comma-separated). Fallback to permissive during MVP.
-env_origins = os.getenv("CORS_ALLOW_ORIGINS", "*")
-origins = [o.strip() for o in env_origins.split(",") if o.strip()]
 app.add_middleware(
     CORSMiddleware,
-    allow_origins=origins,
+    allow_origins=CORS_ALLOW_ORIGINS or ["*"],
     allow_credentials=True,
     allow_methods=["*"],
     allow_headers=["*"],
 )
-
-# Google GenAI client
-_client: Optional[genai.Client] = None
-
-def get_client() -> genai.Client:
-    global _client
-    if _client is None:
-        if not API_KEY:
-            raise RuntimeError("GOOGLE_API_KEY env var is required")
-        _client = genai.Client(api_key=API_KEY)
-    return _client
-
-# --- Garment type classification (top | bottom | full) ---
-_garment_type_cache: dict[str, tuple[float, dict[str, Any]]] = {}
-
-
-def _hash_bytes(data: bytes) -> str:
-    import hashlib
-
-    h = hashlib.sha256()
-    h.update(data)
-    return h.hexdigest()
-
-
-def _first_inline_image_bytes(response: Any) -> bytes | None:
-    """Return the first inline image payload from a Gemini response."""
-
-    for candidate in getattr(response, "candidates", []) or []:
-        content = getattr(candidate, "content", None)
-        parts = getattr(content, "parts", None) if content is not None else None
-        if not parts:
-            continue
-        for part in parts:
-            inline_data = getattr(part, "inline_data", None)
-            data = getattr(inline_data, "data", None) if inline_data is not None else None
-            if data:
-                return data
-    return None
-
-
-def _cache_key(image_hash: str) -> str:
-    return f"{GARMENT_TYPE_CACHE_PREFIX}:{GARMENT_TYPE_CACHE_VERSION}:{image_hash}"
-
-
-def _lock_key(image_hash: str) -> str:
-    return f"{GARMENT_TYPE_CACHE_PREFIX}:lock:{GARMENT_TYPE_CACHE_VERSION}:{image_hash}"
-
-
-def _build_cache_payload(garment_type: str, *, origin: str, ts: float) -> dict[str, Any]:
-    return {
-        "type": garment_type,
-        "origin": origin,
-        "ts": ts,
-        "model_id": MODEL,
-    }
-
-
-async def _get_local_singleflight_lock(image_hash: str) -> asyncio.Lock:
-    async with _local_singleflight_guard:
-        lock = _local_singleflight_locks.get(image_hash)
-        if lock is None:
-            lock = asyncio.Lock()
-            _local_singleflight_locks[image_hash] = lock
-        return lock
-
-
-async def _redis_execute(fn):
-    attempts = max(1, REDIS_OPERATION_RETRIES + 1)
-    last_exc: Exception | None = None
-    for attempt in range(attempts):
-        try:
-            return await asyncio.wait_for(fn(), timeout=REDIS_OP_TIMEOUT_SECONDS)
-        except (asyncio.TimeoutError, Exception) as exc:
-            last_exc = exc
-            if attempt >= attempts - 1:
-                raise
-            await asyncio.sleep(min(0.2, 0.05 * (attempt + 1)))
-    if last_exc is not None:
-        raise last_exc
-    return None
-
-
-def _decode_cache_payload(raw: Any) -> dict[str, Any] | None:
-    if raw is None:
-        return None
-    if isinstance(raw, (bytes, bytearray)):
-        raw_str = raw.decode("utf-8", "ignore")
-    else:
-        raw_str = str(raw)
-    try:
-        payload = json.loads(raw_str)
-    except Exception:
-        return None
-    if not isinstance(payload, dict):
-        return None
-    garment_type = payload.get("type")
-    if not isinstance(garment_type, str):
-        return None
-    normalized = _normalize_garment_type(garment_type)
-    if not normalized:
-        return None
-    payload["type"] = normalized
-    return payload
-
-
-async def _redis_get_payload(redis_client: Any, redis_key: str) -> dict[str, Any] | None:
-    raw = await _redis_execute(lambda: redis_client.get(redis_key))
-    return _decode_cache_payload(raw)
-
-
-async def _redis_set_payload(redis_client: Any, redis_key: str, payload: dict[str, Any]) -> None:
-    data = json.dumps(payload, ensure_ascii=True, separators=(",", ":"))
-    await _redis_execute(lambda: redis_client.set(redis_key, data, ex=GARMENT_TYPE_TTL_SECONDS))
-
-
-async def _acquire_redis_lock(redis_client: Any, lock_key: str, token: str) -> bool:
-    result = await _redis_execute(
-        lambda: redis_client.set(lock_key, token, nx=True, ex=GARMENT_TYPE_LOCK_TTL_SECONDS)
-    )
-    return bool(result)
-
-
-async def _release_redis_lock(redis_client: Any, lock_key: str, token: str) -> None:
-    script = "if redis.call('get', KEYS[1]) == ARGV[1] then return redis.call('del', KEYS[1]) else return 0 end"
-    try:
-        await _redis_execute(lambda: redis_client.eval(script, 1, lock_key, token))
-    except Exception:
-        # Release failures should not block request handling
-        pass
-
-
-async def _wait_for_redis_payload(
-    redis_client: Any, redis_key: str, lock_key: str, *, deadline: float
-) -> dict[str, Any] | None:
-    loop = asyncio.get_running_loop()
-    while loop.time() < deadline:
-        try:
-            payload = await _redis_get_payload(redis_client, redis_key)
-        except Exception as exc:
-            await _record_redis_failure(exc)
-            return None
-        if payload:
-            return payload
-        try:
-            lock_exists = await _redis_execute(lambda: redis_client.exists(lock_key))
-        except Exception as exc:
-            await _record_redis_failure(exc)
-            return None
-        if not lock_exists:
-            # Another worker released lock without writing a value
-            try:
-                payload = await _redis_get_payload(redis_client, redis_key)
-            except Exception as exc:
-                await _record_redis_failure(exc)
-                return None
-            return payload
-        await asyncio.sleep(0.1)
-    return None
-
-def _normalize_garment_type(label: str) -> Optional[str]:
-    s = (label or "").strip().lower()
-    if s in ("top", "bottom", "full"):
-        return s
-    # Heuristic keyword mapping for occasional non-strict replies
-    if any(k in s for k in ("dress", "jumpsuit", "romper", "boilersuit", "overalls", "pinafore", "catsuit", "unitard", "one-piece", "one piece")):
-        return "full"
-    if any(k in s for k in ("jeans", "pants", "trousers", "shorts", "skirt", "leggings", "bottom")):
-        return "bottom"
-    if any(k in s for k in ("t-shirt", "tshirt", "shirt", "blouse", "sweater", "jumper", "hoodie", "cardigan", "jacket", "coat", "vest", "bodysuit", "top")):
-        return "top"
-    return None
-
-async def _classify_garment_type(image_png: bytes, override: Optional[str] = None) -> str:
-    """Classify garment coverage. Returns one of: top|bottom|full.
-
-    Uses an in-memory TTL cache with optional Redis distribution. The cache key is a
-    SHA256 digest of the normalized garment image bytes.
-    """
-
-    normalized_override = _normalize_garment_type(override or "") if override else None
-    if normalized_override:
-        return normalized_override
-    if not GARMENT_TYPE_CLASSIFY:
-        return "full"
-
-    loop = asyncio.get_running_loop()
-    image_hash = _hash_bytes(image_png)
-    now = loop.time()
-    cached_entry = _garment_type_cache.get(image_hash)
-    if cached_entry and cached_entry[0] > now:
-        return cached_entry[1]["type"]
-
-    redis_client = await get_redis_client()
-    redis_key = _cache_key(image_hash)
-    lock_key = _lock_key(image_hash)
-    if redis_client and GARMENT_TYPE_TTL_SECONDS > 0:
-        try:
-            payload = await _redis_get_payload(redis_client, redis_key)
-        except Exception as exc:  # pragma: no cover - network errors
-            await _record_redis_failure(exc)
-            redis_client = None
-        else:
-            if payload:
-                _garment_type_cache[image_hash] = (
-                    loop.time() + GARMENT_TYPE_TTL_SECONDS,
-                    payload,
-                )
-                return payload["type"]
-
-    lock_token = uuid.uuid4().hex
-    redis_lock_acquired = False
-    lock_client: Any | None = None
-    local_lock: asyncio.Lock | None = None
-    local_lock_acquired = False
-    try:
-        if redis_client:
-            try:
-                redis_lock_acquired = await _acquire_redis_lock(redis_client, lock_key, lock_token)
-            except Exception as exc:  # pragma: no cover - network errors
-                await _record_redis_failure(exc)
-                redis_client = None
-            else:
-                if not redis_lock_acquired:
-                    deadline = loop.time() + GARMENT_TYPE_LOCK_WAIT_SECONDS
-                    payload = await _wait_for_redis_payload(
-                        redis_client, redis_key, lock_key, deadline=deadline
-                    )
-                    if payload:
-                        if GARMENT_TYPE_TTL_SECONDS > 0:
-                            _garment_type_cache[image_hash] = (
-                                loop.time() + GARMENT_TYPE_TTL_SECONDS,
-                                payload,
-                            )
-                        return payload["type"]
-                else:
-                    lock_client = redis_client
-        if not redis_lock_acquired:
-            local_lock = await _get_local_singleflight_lock(image_hash)
-            await local_lock.acquire()
-            local_lock_acquired = True
-
-        now = loop.time()
-        cached_entry = _garment_type_cache.get(image_hash)
-        if cached_entry and cached_entry[0] > now:
-            return cached_entry[1]["type"]
-
-        if redis_client:
-            try:
-                payload = await _redis_get_payload(redis_client, redis_key)
-            except Exception as exc:  # pragma: no cover - network errors
-                await _record_redis_failure(exc)
-                redis_client = None
-            else:
-                if payload:
-                    if GARMENT_TYPE_TTL_SECONDS > 0:
-                        _garment_type_cache[image_hash] = (
-                            loop.time() + GARMENT_TYPE_TTL_SECONDS,
-                            payload,
-                        )
-                    return payload["type"]
-
-        instruction = (
-            "From the attached garment image, classify coverage for try-on. "
-            "Return ONLY one word: top (upper body), bottom (lower body), or full (one piece covering upper+lower like dress/jumpsuit/romper/overalls). "
-            "Output: top|bottom|full."
-        )
-        parts = [
-            types.Part.from_text(text=instruction),
-            types.Part.from_bytes(data=image_png, mime_type="image/png"),
-        ]
-        try:
-            resp = await _genai_generate_with_retries(parts, attempts=2)
-            label_text: Optional[str] = None
-            for c in getattr(resp, "candidates", []) or []:
-                content = getattr(c, "content", None)
-                prts = getattr(content, "parts", None) if content is not None else None
-                if not prts:
-                    continue
-                for p in prts:
-                    if getattr(p, "text", None):
-                        label_text = p.text
-                        break
-                if label_text:
-                    break
-            t = _normalize_garment_type(label_text or "")
-            garment_type = t or "full"
-        except Exception:
-            garment_type = "full"
-        payload = _build_cache_payload(garment_type, origin="classifier", ts=time.time())
-        if GARMENT_TYPE_TTL_SECONDS > 0:
-            _garment_type_cache[image_hash] = (
-                loop.time() + GARMENT_TYPE_TTL_SECONDS,
-                payload,
-            )
-        if redis_client and GARMENT_TYPE_TTL_SECONDS > 0:
-            try:
-                await _redis_set_payload(redis_client, redis_key, payload)
-            except Exception as exc:  # pragma: no cover - network errors
-                await _record_redis_failure(exc)
-        return garment_type
-    finally:
-        if local_lock_acquired and local_lock is not None and local_lock.locked():
-            local_lock.release()
-        if lock_client is not None and redis_lock_acquired:
-            try:
-                await _release_redis_lock(lock_client, lock_key, lock_token)
-            except Exception:
-                pass
-
-
-
-def _normalize_to_png_limited(raw_bytes: bytes, *, max_px: int = 2048) -> bytes:
-    """Decode bytes with PIL, optionally downscale to keep within max_px on longer side,
-    and re-encode as PNG. Returns PNG bytes.
-    """
-    src = Image.open(BytesIO(raw_bytes))
-    try:
-        src = src.convert("RGBA")
-        w, h = src.size
-        if max(w, h) > max_px:
-            # Preserve aspect ratio; downscale in-place
-            scale = max_px / float(max(w, h))
-            new_size = (int(w * scale), int(h * scale))
-            src = src.resize(new_size, Image.LANCZOS)
-        out = BytesIO()
-        src.save(out, format="PNG")
-        out.seek(0)
-        return out.getvalue()
-    finally:
-        try:
-            src.close()
-        except Exception:
-            pass
-
-
-async def _genai_generate_with_retries(parts: list[types.Part], *, attempts: int = 2):
-    """Call GenAI with short retries for transient 5xx/429 errors.
-
-    Returns the raw response object.
-    """
-    last_exc: Exception | None = None
-    for i in range(max(1, attempts)):
-        try:
-            return await asyncio.to_thread(
-                get_client().models.generate_content,
-                model=MODEL,
-                contents=types.Content(role="user", parts=parts),
-            )
-        except genai_errors.APIError as e:
-            last_exc = e
-            # Retry on server/internal or quota/too-many-requests
-            code = getattr(e, "code", None)
-            msg = (getattr(e, "message", "") or "").lower()
-            if code in (500, 502, 503) or "internal" in msg or code == 429:
-                await asyncio.sleep(0.6 + 0.4 * i)
-                continue
-            raise
-    # After attempts exhausted, re-raise last
-    assert last_exc is not None
-    raise last_exc
-
-
-@app.get("/health")
-async def health():
-    return {"ok": True, "model": MODEL}
+app.include_router(api_router)
 
 
 @app.on_event("startup")
-async def on_startup():
-    # Fail fast if DB cannot be initialized so the service doesn't run half‑broken
+async def on_startup() -> None:
     try:
         await init_db()
-        logger.info("DB initialized")
+        LOGGER.info("DB initialized")
     except Exception:
-        logger.exception("Failed to initialize DB (startup)")
-        # Re-raise to crash the worker/container; orchestration should restart once env is fixed
+        LOGGER.exception("Failed to initialize DB (startup)")
         raise
     if REDIS_URL and redis_asyncio is not None:
         try:
             client = await get_redis_client()
             if client is not None:
-                logger.info("Redis client ready")
+                LOGGER.info("Redis client ready")
         except Exception:
-            logger.exception("Redis startup check failed")
+            LOGGER.exception("Redis startup check failed")
 
 
 @app.on_event("shutdown")
-async def on_shutdown():
-    global _redis_client
-    client = _redis_client
-    _redis_client = None
-    if client is not None:
-        try:
-            await client.close()
-        except Exception:
-            pass
-
-
-def _require_admin(authorization: str | None) -> None:
-    token = os.getenv("ADMIN_BEARER_TOKEN")
-    if not token:
-        raise HTTPException(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail="ADMIN_BEARER_TOKEN not configured")
-    if not authorization or not authorization.startswith("Bearer "):
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Missing bearer token")
-    if authorization.split(" ", 1)[1] != token:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Forbidden")
-
-
-@app.post("/admin/init-db")
-async def admin_init_db(authorization: str | None = Header(default=None, alias="Authorization")):
-    """Admin-only endpoint to (re)create all SQL tables using SQLAlchemy metadata.
-
-    Useful after resetting the database. Protected by ADMIN_BEARER_TOKEN.
-    """
-    _require_admin(authorization)
-    try:
-        await init_db()
-        return {"ok": True, "message": "DB initialized"}
-    except Exception as e:
-        logger.exception("Admin init-db failed")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-def _normalize_choice(value: str, allowed: list[str], default: str) -> str:
-    value = (value or "").strip().lower()
-    return value if value in allowed else default
-
-
-def build_env_prompt(user_prompt: Optional[str] = None) -> str:
-    """Build the Studio Environment generation instruction.
-
-    Goals
-    - Always a mirror scene. Use the attached source image as the scene seed.
-    - Produce a new, photorealistic environment that is coherent with the source reflection.
-    - Randomize tastefully while keeping the mirror's size/placement consistent.
-    """
-    def q(s: Optional[str]) -> str:
-        return (s or "").strip()
-
-    lines: list[str] = []
-    lines.append("TASK")
-    lines.append(
-        "Generate a new photorealistic mirror environment image for future garment try-ons. "
-        "All scenes are mirror reflections. Use the attached image as a seed reference."
-    )
-    lines.append("")
-    lines.append("HARD CONSTRAINTS")
-    lines.append("- No people; no text; no logos or brand marks.")
-    lines.append("- Useful as a backdrop: good, even lighting; natural shadows; mild depth of field.")
-    lines.append("- Keep an empty, clean mid-ground area in the mirror where a person could plausibly stand.")
-    lines.append(
-        "- Mirror coherence: preserve the mirror opening size, aspect ratio, and on-image placement; keep the camera viewpoint and reflection geometry plausible."
-    )
-    lines.append("")
-    lines.append("SCENE CONSISTENCY (interior vs exterior)")
-    lines.append(
-        "- Deduce the scene category and subtype solely from the attached source reflection. Recreate a fresh variant of the SAME subtype (e.g., bedroom → bedroom), or if ambiguous, keep it plausibly neutral while preserving interior vs exterior. Do NOT switch interior ↔ exterior and do NOT invent a different subtype."
-    )
-    lines.append("")
-    lines.append("MIRROR FRAME BEHAVIOR")
-    lines.append(
-        "- Randomize the mirror frame style (material, color, ornamentation) while keeping the mirror's opening size and aspect ratio IDENTICAL and its on-image position UNCHANGED. Do not scale, stretch, move, or tilt the opening."
-    )
-    lines.append("")
-    lines.append("RANDOMIZATION")
-    lines.append("- Vary room subtype, color palette, materials, and tasteful props.")
-    lines.append("- Keep background clutter low; avoid busy textures near the mid torso area.")
-    lines.append("- Maintain photorealism; avoid stylization.")
-    lines.append("")
-    if q(user_prompt):
-        lines.append("USER WISHES")
-        lines.append(f"\"{q(user_prompt)}\"")
-        lines.append(
-            "Apply ONLY if consistent with photorealism, cleanliness, mirror-reflection setup, and usability as a try-on backdrop. "
-            "Do not change the scene category (interior vs exterior) and do not change the mirror size/placement."
-        )
-        lines.append("")
-    lines.append("NEGATIVE GUIDANCE")
-    lines.append(
-        "cluttered, illegible text, signage, extreme wide-angle distortion, heavy motion blur, horror/abandoned spaces, AI artifacts"
-    )
-    return "\n".join(lines)
-
-
-def build_mirror_selfie_prompt(
-    *,
-    gender: str,
-    environment: str,
-    pose: str,
-    extra: str,
-    use_env_image: bool,
-    use_person_image: bool,
-    person_description: Optional[str] = None,
-) -> str:
-    """Build the core prompt (image edit) using conditioned controls and strict constraints.
-
-    - If an environment ref is provided, avoid prescribing a textual environment.
-    - Optionally include a short person description when no person image is provided.
-    - Mirror-selfie style retained for consistency with the app UI.
-    """
-    # Normalize fields and ensure empty strings for missing conditioned controls
-    def norm(x: Optional[str]) -> str:
-        return (x or "").strip()
-
-    conditioned_gender = norm(gender)
-    conditioned_env = norm(environment) if not use_env_image else ""
-    conditioned_pose = norm(pose)
-    conditioned_extra = norm(extra)
-
-    lines: list[str] = []
-    lines.append("High-level goals")
-    lines.append("- Prioritize faithful garment reproduction (shape, fabric, color, print, logos).")
-    lines.append("- Natural, flattering person and pose; avoid distortions.")
-    lines.append("- Use environment reference if provided; otherwise synthesize the requested environment.")
-    lines.append("- Keep results PG-13 (no nudity; no explicit content).")
-    lines.append("")
-    lines.append("TASK")
-    task_parts: list[str] = []
-    task_parts.append("You render a photorealistic mirror selfie of a person wearing the provided garment.")
-    if use_person_image:
-        task_parts.append("Use the attached person reference image; keep identity cues, hair, and overall build consistent (the face may be occluded by the phone).")
-    elif person_description:
-        task_parts.append("No person image; use the provided person description to guide identity (the face may be occluded by the phone).")
-    else:
-        task_parts.append("No person reference; synthesize a plausible model matching the selected person.")
-    if use_env_image:
-        task_parts.append("Use the attached environment reference as a mirror scene; match its lighting, camera angle, color palette, and depth of field.")
-    else:
-        task_parts.append("No environment reference; synthesize a clean mirror setting consistent with the requested environment.")
-    task_parts.append("Keep an amateur smartphone look.")
-    lines.append(" ".join(task_parts))
-    lines.append("")
-    lines.append("REQUIRED OUTPUT")
-    lines.append("- One 2D PNG photo, realistic lighting and skin.")
-    lines.append("- The garment must be the dominant subject, clearly visible and not occluded.")
-    lines.append("")
-    lines.append("HARD CONSTRAINTS (must follow)")
-    lines.append("1) Garment fidelity: keep exact silhouette, color, fabric texture, print scale/alignment, closures, and logos from the garment image; avoid moiré.")
-    lines.append("2) Body realism: natural proportions; hands with five distinct fingers; no merges/extra digits; no warped limbs.")
-    lines.append("3) Face realism: plausible expression; no duplicates/melting; preserve identity cues if a person ref is provided.")
-    lines.append("4) Clothing fit: believable size and drape; respect gravity and fabric stiffness.")
-    lines.append("5) Clean output: no watermarks, no AI artifacts, no text overlays, no added/brand-new logos.")
-    lines.append("6) Safety: PG-13; no explicit content.")
-    lines.append("7) Mirror selfie: keep a smartphone-in-mirror aesthetic; if the phone occludes the face, do so consistently (with correct reflection) without hiding key garment details.")
-    lines.append("8) Garment usage: the person must be wearing the uploaded garment; do not omit or replace it.")
-    lines.append("")
-    lines.append("CONDITIONED CONTROLS")
-    # Show person label instead of gender; omit value when a person image is attached
-    lines.append(f"- Person: {conditioned_gender if (conditioned_gender and not use_person_image) else '""'}")
-    # Rename Environment to Scene; omit value when an env image is attached (conditioned_env is empty in that case)
-    lines.append(f"- Scene: {conditioned_env if conditioned_env else '""'}")
-    lines.append(f"- Pose: {conditioned_pose if conditioned_pose else '""'}")
-    lines.append(f"- Notes: \"{conditioned_extra.replace('\\n', ' ')}\"")
-    lines.append("")
-    lines.append("STYLE & CAMERA DIRECTION")
-    lines.append("- Smartphone mirror-selfie aesthetic; natural colors; mild grain acceptable.")
-    lines.append("- Shot type: 3/4 body by default unless the pose implies otherwise (ensure garment fully visible).")
-    lines.append("- Camera: 35–70mm equivalent perspective; natural lens distortion; f/2.8–f/5.6; soft bokeh if indoors.")
-    lines.append("- Lighting: match environment reference if given; otherwise soft directional key + gentle fill; mild rim light for separation.")
-    lines.append("- Composition: center subject in mirror; show phone and hand; avoid cropping garment edges; keep hands visible when natural.")
-    lines.append("")
-    lines.append("SCENE BEHAVIOR")
-    if use_env_image:
-        lines.append("- Use the attached environment reference as a mirror scene; imitate its scene category, palette, light direction, shadows, and depth of field; keep any mirror frame consistent.")
-    else:
-        lines.append("- No environment reference: synthesize a clean mirror setting that complements the garment; minimal, elegant background; avoid busy textures behind the torso.")
-    lines.append("")
-    lines.append("PERSON BEHAVIOR")
-    if use_person_image:
-        lines.append("- Person reference: use the attached image; keep face, hair, skin tone, and general build consistent (face may be partly occluded by phone).")
-    else:
-        lines.append("- No person reference: synthesize a plausible model; natural expression.")
-        if person_description:
-            lines.append("- Use a person that matches this description.")
-            lines.append(f"- Person description: {person_description}")
-    lines.append("- Hand pose: holding a smartphone naturally; fingers look correct; phone and reflection visible.")
-    lines.append("")
-    lines.append("POSE RENDERING")
-    lines.append(f"- Enforce the requested pose: {conditioned_pose if conditioned_pose else '""'}. Make it balanced and anatomically plausible.")
-    lines.append("- Ensure the garment remains fully visible and not occluded by the phone or pose; if sitting or lying down, do not let the pose hide key garment areas.")
-    lines.append("")
-    lines.append("QUALITY CHECK BEFORE OUTPUT")
-    lines.append("- Fingers: five per hand; shapes correct; no extra/merged digits.")
-    lines.append("- Garment: crisp edges; seams/hemlines visible; prints/logos accurate.")
-    lines.append("- Face: no duplicates; no melting; if visible, eyes symmetrical; otherwise occluded by phone.")
-    lines.append("- Mirror: phone and reflection consistent; no extra phones; no camera artifacts.")
-    lines.append("- Background: clean and coherent; matches env ref if provided.")
-    lines.append("")
-    lines.append("NEGATIVE GUIDANCE (avoid)")
-    lines.append("blurry, over-saturated, HDR halos, duplicated limbs, extra fingers, merged fingers, warped faces, melted textures, text overlays, watermarks, added/brand-new logos, heavy beauty retouching, studio glamour look, ring-light glare, tripod/DSLR look, explicit content, busy background patterns near the torso.")
-    lines.append("")
-    lines.append("END OF INSTRUCTIONS")
-
-    return "\n".join(lines)
-
-
-@app.post("/edit")
-async def edit(
-    image: UploadFile = File(...),
-    gender: str = Form("woman"),
-    environment: str = Form("studio"),
-    poses: list[str] = Form(None),
-    extra: str = Form(""),
-    env_default_s3_key: str | None = Form(None),
-    model_default_s3_key: str | None = Form(None),
-    model_description_text: str | None = Form(None),
-    prompt_override: str | None = Form(None),
-    garment_type_override: str | None = Form(None),
-    x_user_id: str | None = Header(default=None, alias="X-User-Id"),
-):
-    try:
-        if not image or not image.filename:
-            return JSONResponse({"error": "image file required"}, status_code=400)
-        # Read uploaded image bytes, normalize to PNG and downscale if needed
-        raw_bytes = await image.read()
-        if len(raw_bytes) > 20 * 1024 * 1024:
-            return JSONResponse({"error": "image too large (max ~20MB)"}, status_code=413)
-        try:
-            png_bytes = _normalize_to_png_limited(raw_bytes, max_px=2048)
-        except Exception:
-            return JSONResponse({"error": "invalid or unsupported image format"}, status_code=400)
-
-        # Validate and normalize options
-        gender = _normalize_choice(gender, ["woman", "man"], "woman")
-        environment = _normalize_choice(environment, ["studio", "street", "bed", "beach", "indoor"], "studio")
-        # Normalize poses array (multi-select). Accept up to 3 unique values
-        allowed_poses = ["standing", "sitting", "lying down", "walking"]
-        if not poses:
-            poses = []
-        if not isinstance(poses, list):
-            poses = [poses]
-        norm_poses: list[str] = []
-        for p in poses:
-            p_norm = _normalize_choice(p, allowed_poses, "")
-            if p_norm and p_norm not in norm_poses:
-                norm_poses.append(p_norm)
-            if len(norm_poses) >= 3:
-                break
-        extra = (extra or "").strip()
-        if len(extra) > 200:
-            extra = extra[:200]
-
-        if not norm_poses:
-            norm_poses = ["standing"]
-
-        # Classify garment type (with optional override)
-        garment_type = await _classify_garment_type(png_bytes, garment_type_override)
-
-        # Build prompt and parts in numbered image order: 1=garment, 2=person (opt), 3=environment (opt)
-        use_env_image = bool(env_default_s3_key)
-        use_person_image = bool(model_default_s3_key)
-        pose_str = norm_poses[0] if norm_poses else ""
-        prompt_variant = "detailed"
-        if prompt_override and prompt_override.strip():
-            prompt_text = prompt_override.strip()
-        else:
-            prompt_text = classic_detailed(
-                gender=gender,
-                environment=environment,
-                pose=pose_str,
-                use_person_image=use_person_image,
-                use_env_image=use_env_image,
-                person_description=(model_description_text if (model_description_text and not use_person_image) else None),
-                garment_type=garment_type,
-            )
-        parts: list[types.Part] = [types.Part.from_text(text=prompt_text)]
-        if (not use_person_image) and model_description_text:
-            parts.append(types.Part.from_text(text=f"Person description: {model_description_text}"))
-        # Image 1: garment (uploaded)
-        parts.append(types.Part.from_bytes(data=png_bytes, mime_type="image/png"))
-        person_key_used: str | None = None
-        env_key_used: str | None = None
-        # Image 2: person (optional)
-        if model_default_s3_key:
-            try:
-                person_bytes, person_mime = get_object_bytes(model_default_s3_key)
-                parts.append(types.Part.from_bytes(data=person_bytes, mime_type=person_mime or "image/png"))
-                person_key_used = model_default_s3_key
-            except Exception:
-                person_key_used = None
-        # Image 3: environment (optional)
-        if env_default_s3_key:
-            try:
-                env_bytes, env_mime = get_object_bytes(env_default_s3_key)
-                parts.append(types.Part.from_bytes(data=env_bytes, mime_type=env_mime or "image/png"))
-                env_key_used = env_default_s3_key
-            except Exception:
-                env_key_used = None
-
-        resp = await _genai_generate_with_retries(parts, attempts=2)
-        png_bytes = _first_inline_image_bytes(resp)
-        if png_bytes:
-            # Upload to S3
-            bucket, key = upload_image(png_bytes, pose=norm_poses[0])
-            # Persist to DB
-            async with db_session() as session:
-                rec = Generation(
-                    s3_key=key,
-                    pose=norm_poses[0],
-                    prompt=prompt_text,
-                    options_json={
-                        "gender": gender,
-                        "environment": environment,
-                        "poses": norm_poses,
-                        "extra": extra,
-                        "env_default_s3_key": env_key_used,
-                        "model_default_s3_key": person_key_used,
-                        "model_description_text": (model_description_text if not person_key_used else None),
-                        "garment_type": garment_type,
-                        "garment_type_override": (garment_type_override if garment_type_override else None),
-                        "user_id": x_user_id,
-                        "prompt_variant": prompt_variant,
-                    },
-                    model=MODEL,
-                )
-                session.add(rec)
-            # Stream bytes back for current UI
-            return StreamingResponse(BytesIO(png_bytes), media_type="image/png")
-        # Fallback to concise variant if no image and no explicit override
-        if not (prompt_override and prompt_override.strip()):
-            try:
-                prompt_variant = "concise"
-                prompt_text = classic_concise(
-                    gender=gender,
-                    environment=environment,
-                    pose=pose_str,
-                    use_person_image=use_person_image,
-                    use_env_image=use_env_image,
-                    person_description=(model_description_text if (model_description_text and not use_person_image) else None),
-                    garment_type=garment_type,
-                )
-                parts[0] = types.Part.from_text(text=prompt_text)
-                resp2 = await _genai_generate_with_retries(parts, attempts=1)
-                png_bytes2 = _first_inline_image_bytes(resp2)
-                if png_bytes2:
-                    bucket, key = upload_image(png_bytes2, pose=norm_poses[0])
-                    async with db_session() as session:
-                        rec = Generation(
-                            s3_key=key,
-                            pose=norm_poses[0],
-                            prompt=prompt_text,
-                            options_json={
-                                "gender": gender,
-                                "environment": environment,
-                                "poses": norm_poses,
-                                "extra": extra,
-                                "env_default_s3_key": env_key_used,
-                                "model_default_s3_key": person_key_used,
-                                "model_description_text": (model_description_text if not person_key_used else None),
-                                "garment_type": garment_type,
-                                "garment_type_override": (garment_type_override if garment_type_override else None),
-                                "user_id": x_user_id,
-                                "prompt_variant": prompt_variant,
-                            },
-                            model=MODEL,
-                        )
-                        session.add(rec)
-                    return StreamingResponse(BytesIO(png_bytes2), media_type="image/png")
-            except Exception:
-                pass
-        try:
-            cand_count = len(getattr(resp, "candidates", []) or [])
-        except Exception:
-            cand_count = -1
-        logger.warning(
-            "edit: no image from model (candidates=%s, prompt_len=%s, use_env=%s, use_person=%s)",
-            cand_count,
-            len(prompt_text or ""),
-            bool(env_default_s3_key),
-            bool(model_default_s3_key),
-        )
-        return JSONResponse({"error": "no edited image from model"}, status_code=502)
-    except genai_errors.APIError as e:
-        logger.exception("GenAI API error on /edit")
-        return JSONResponse({"error": e.message, "code": e.code}, status_code=502)
-    except Exception as e:
-        logger.exception("Unhandled error on /edit")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-# --- Environment sources and generation ---
-
-@app.post("/env/sources/upload")
-async def upload_env_sources(files: list[UploadFile] = File(...)):
-    try:
-        stored = []
-        for f in files:
-            data = await f.read()
-            bucket, key = upload_source_image(data, mime=f.content_type)
-            async with db_session() as session:
-                rec = EnvSource(s3_key=key)
-                session.add(rec)
-            stored.append({"s3_key": key})
-        return {"ok": True, "count": len(stored), "items": stored}
-    except Exception as e:
-        logger.exception("Failed to upload env sources")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.get("/env/sources")
-async def list_env_sources():
-    try:
-        async with db_session() as session:
-            stmt = select(EnvSource.s3_key).order_by(EnvSource.created_at.desc())
-            res = await session.execute(stmt)
-            items = [row[0] for row in res.all()]
-        return {"ok": True, "count": len(items), "items": items}
-    except Exception as e:
-        logger.exception("Failed to list env sources")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.delete("/env/sources")
-async def delete_env_sources():
-    try:
-        # Fetch all keys
-        async with db_session() as session:
-            stmt = select(EnvSource.s3_key)
-            res = await session.execute(stmt)
-            keys = [row[0] for row in res.all()]
-        # Delete from S3 first
-        delete_objects(keys)
-        # Delete from DB
-        async with db_session() as session:
-            await session.execute(text("DELETE FROM env_sources"))
-        return {"ok": True, "deleted": len(keys)}
-    except Exception as e:
-        logger.exception("Failed to delete env sources")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-async def _generate_env_with_random_source(prompt_text: str, *, options: dict[str, Any]):
-    """Shared helper for /env/random and /env/generate flows."""
-
-    async with db_session() as session:
-        stmt = text("SELECT s3_key FROM env_sources ORDER BY RANDOM() LIMIT 1")
-        res = await session.execute(stmt)
-        row = res.first()
-    if not row:
-        return JSONResponse({"error": "no sources uploaded"}, status_code=400)
-
-    source_key = row[0]
-    src_bytes, mime = get_object_bytes(source_key)
-    parts = [
-        types.Part.from_text(text=prompt_text),
-        types.Part.from_bytes(data=src_bytes, mime_type=mime),
-    ]
-    resp = await _genai_generate_with_retries(parts, attempts=2)
-    png_bytes = _first_inline_image_bytes(resp)
-    if not png_bytes:
-        return JSONResponse({"error": "no image from model"}, status_code=502)
-
-    _, key = upload_image(png_bytes, pose="env")
-    payload = dict(options)
-    payload["source_s3_key"] = source_key
-    async with db_session() as session:
-        rec = Generation(
-            s3_key=key,
-            pose="env",
-            prompt=prompt_text,
-            options_json=payload,
-            model=MODEL,
-        )
-        session.add(rec)
-    return StreamingResponse(BytesIO(png_bytes), media_type="image/png")
-
-
-@app.post("/env/random")
-async def generate_env_random(x_user_id: str | None = Header(default=None, alias="X-User-Id")):
-    """Pick a random stored source and generate with strict instruction."""
-    try:
-        instruction = build_env_prompt()
-        return await _generate_env_with_random_source(
-            instruction,
-            options={"mode": "random", "user_id": x_user_id},
-        )
-    except Exception as e:
-        logger.exception("env random failed")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.post("/env/generate")
-async def generate_env(prompt: str = Form(""), x_user_id: str | None = Header(default=None, alias="X-User-Id")):
-    try:
-        full = build_env_prompt(prompt)
-        user_prompt = (prompt or "").strip()
-        return await _generate_env_with_random_source(
-            full,
-            options={
-                "mode": "prompt",
-                "user_prompt": user_prompt,
-                "user_id": x_user_id,
-            },
-        )
-    except Exception as e:
-        logger.exception("env generate failed")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.get("/env/generated")
-async def list_generated(x_user_id: str | None = Header(default=None, alias="X-User-Id")):
-    try:
-        async with db_session() as session:
-            if x_user_id:
-                stmt = (
-                    select(Generation.s3_key, Generation.created_at)
-                    .where(Generation.pose == "env")
-                    .where(text("(options_json->>'user_id') = :uid")).params(uid=x_user_id)
-                    .order_by(Generation.created_at.desc())
-                    .limit(200)
-                )
-            else:
-                # No user id provided: return empty list to avoid cross-user leakage
-                stmt = select(Generation.s3_key, Generation.created_at).where(text("1=0"))
-            res = await session.execute(stmt)
-            rows = res.all()
-            items = []
-            for row in rows:
-                key = row[0]
-                created = row[1]
-                try:
-                    url = generate_presigned_get_url(key)
-                except Exception:
-                    url = None
-                items.append({
-                    "s3_key": key,
-                    "created_at": created.isoformat(),
-                    "url": url,
-                })
-        return {"ok": True, "count": len(items), "items": items}
-    except Exception as e:
-        logger.exception("Failed to list generated images")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.get("/env/image")
-async def get_generated_image(s3_key: str):
-    try:
-        data, content_type = get_object_bytes(s3_key)
-        return StreamingResponse(BytesIO(data), media_type=content_type)
-    except Exception as e:
-        logger.exception("Failed to fetch generated image")
-        return JSONResponse({"error": str(e)}, status_code=404)
-
-
-# --- Defaults (select up to 5, name them) ---
-
-@app.get("/env/defaults")
-async def list_defaults(x_user_id: str | None = Header(default=None, alias="X-User-Id")):
-    try:
-        async with db_session() as session:
-            if not x_user_id:
-                rows = []
-            else:
-                stmt = (
-                    select(EnvDefaultUser.s3_key, EnvDefaultUser.name)
-                    .where(EnvDefaultUser.user_id == x_user_id)
-                    .order_by(EnvDefaultUser.created_at.desc())
-                )
-                res = await session.execute(stmt)
-                rows = res.all()
-            items = []
-            for row in rows:
-                key, name = row
-                try:
-                    url = generate_presigned_get_url(key)
-                except Exception:
-                    url = None
-                items.append({"s3_key": key, "name": name, "url": url})
-        return {"ok": True, "items": items}
-    except Exception as e:
-        logger.exception("Failed to list defaults")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-# --- Model generation (randomize a person with optional env reference) ---
-
-def _normalize_gender(g: str) -> str:
-    g = (g or "").strip().lower()
-    return g if g in ("man", "woman") else "man"
-
-
-def build_model_prompt(gender: str, user_prompt: Optional[str]) -> str:
-    """Prompt for generating a reusable person model reference.
-
-    - Uses an attached source person image as the reference.
-    - Produce a photorealistic {gender} with the SAME clothing and the SAME background as the source.
-    - Vary only identity cues (face/identity), not the outfit or environment.
-    """
-    def q(s: Optional[str]) -> str:
-        return (s or "").strip()
-
-    lines: list[str] = []
-    lines.append("TASK")
-    lines.append(
-        f"Generate a photorealistic {gender} model portrait/full-body for try-on catalogs. "
-        "Use the attached person image as the reference: keep the SAME clothing and the SAME background; change only the person identity to a different, plausible {gender}."
-    )
-    lines.append("")
-    lines.append("HARD CONSTRAINTS")
-    lines.append("- Natural, friendly expression; neutral makeup (if applicable).")
-    lines.append("- Balanced body proportions; realistic hands.")
-    lines.append("- Clothing: preserve EXACTLY what the source person wears (same garments, colors, materials, prints, logos if present). Do not alter fit or style.")
-    lines.append("- Background/scene: preserve EXACTLY what is in the source (same location, props, lighting, palette, depth of field). Do not replace or restage.")
-    lines.append("- No explicit content; keep PG-13.")
-    lines.append("")
-    lines.append("CAMERA & LIGHT")
-    lines.append("- Preserve the source camera perspective and lighting as part of the unchanged background.")
-    lines.append("- Framing: 3/4 body if plausible from the source; otherwise match source framing.")
-    lines.append("")
-    lines.append("RANDOMIZATION")
-    lines.append("- Randomize identity ONLY (different person of the same gender). DO NOT change clothing or background.")
-    if q(user_prompt):
-        lines.append("")
-        lines.append("USER WISHES")
-        lines.append(f"\"{q(user_prompt)}\"")
-        lines.append("Apply only if consistent with realism and constraints above.")
-    lines.append("")
-    lines.append("NEGATIVE GUIDANCE")
-    lines.append("over-retouched skin, plastic look, extreme stylization, caricature, heavy vignettes, AI artifacts")
-    return "\n".join(lines)
-
-
-@app.post("/model/generate")
-async def model_generate(
-    image: UploadFile | None = File(None),
-    gender: str = Form("man"),
-    prompt: str = Form(""),
-    x_user_id: str | None = Header(default=None, alias="X-User-Id"),
-):
-    try:
-        gender = _normalize_gender(gender)
-        user_prompt = (prompt or "").strip()
-
-        # Build instruction
-        instruction = build_model_prompt(gender, user_prompt if user_prompt else None)
-        parts: list[types.Part] = [types.Part.from_text(text=instruction)]
-
-        # Resolve person source image
-        src_png_bytes: Optional[bytes] = None
-        if image and getattr(image, "filename", None):
-            # Read and normalize uploaded image to PNG bytes
-            raw_bytes = await image.read()
-            if len(raw_bytes) > 10 * 1024 * 1024:
-                return JSONResponse({"error": "image too large (max ~10MB)"}, status_code=413)
-            src = Image.open(BytesIO(raw_bytes))
-            buf = BytesIO()
-            src.convert("RGBA").save(buf, format="PNG")
-            buf.seek(0)
-            src_png_bytes = buf.getvalue()
-            # Persist uploaded source for admin library
-            try:
-                _, src_key = upload_model_source_image(src_png_bytes, gender=gender, mime="image/png")
-                async with db_session() as session:
-                    session.add(ModelSource(gender=gender, s3_key=src_key))
-            except Exception:
-                pass
-        else:
-            # No image provided -> pick the most recently uploaded admin source of this gender
-            async with db_session() as session:
-                stmt = text("SELECT s3_key FROM model_sources WHERE gender = :g ORDER BY created_at DESC LIMIT 1")
-                res = await session.execute(stmt, {"g": gender})
-                row = res.first()
-            if not row:
-                return JSONResponse({"error": f"no model sources uploaded for gender '{gender}'"}, status_code=400)
-            src_bytes, src_mime = get_object_bytes(row[0])
-            src_png_bytes = src_bytes  # assume stored as PNG; mime guards below
-
-        # Attach source as input
-        parts.append(types.Part.from_bytes(data=src_png_bytes, mime_type="image/png"))
-
-        resp = await asyncio.to_thread(
-            get_client().models.generate_content,
-            model=MODEL,
-            contents=types.Content(role="user", parts=parts),
-        )
-        png_bytes = _first_inline_image_bytes(resp)
-        if png_bytes:
-            bucket, key = upload_image(png_bytes, pose=f"model-{gender}")
-            async with db_session() as session:
-                rec = Generation(
-                    s3_key=key,
-                    pose=f"model-{gender}",
-                    prompt=instruction,
-                    options_json={
-                        "mode": "model",
-                        "gender": gender,
-                        "user_prompt": user_prompt,
-                        "user_id": x_user_id,
-                    },
-                    model=MODEL,
-                )
-                session.add(rec)
-            # Follow-up: generate a detailed identity description for this image
-            try:
-                describe_prompt = (
-                    "Describe this person precisely for identity reference (plain text, MINIMUM 500 words). "
-                    "Focus strictly on identity cues, not clothing or background. Include: perceived gender; approximate age range; height impression; build; posture; skin tone with nuance; undertone; face shape; forehead; hairline; hair color; highlights/lowlights; hair length; hair texture; parting; typical styles; eyebrows (shape, thickness, arch); eyes (color, shape, spacing, eyelids); eyelashes; nose (bridge, tip, width); cheeks; lips (shape, fullness, Cupid's bow); chin; jawline; ears; facial hair (if any, density and shape); teeth and smile; notable features (freckles, moles, scars, dimples, birthmarks); accessories (glasses, earrings, piercings). "
-                    "Use neutral, respectful language; avoid judgments; avoid clothing/brand/background mentions; no lists of instructions—write a cohesive, descriptive paragraph or two with at least 500 words."
-                )
-                desc_parts = [
-                    types.Part.from_text(text=describe_prompt),
-                    types.Part.from_bytes(data=png_bytes, mime_type="image/png"),
-                ]
-                desc_resp = get_client().models.generate_content(
-                    model=MODEL,
-                    contents=types.Content(role="user", parts=desc_parts),
-                )
-                description_text = None
-                for dc in getattr(desc_resp, "candidates", []) or []:
-                    dcontent = getattr(dc, "content", None)
-                    dparts = getattr(dcontent, "parts", None) if dcontent is not None else None
-                    if dparts:
-                        for part in dparts:
-                            if getattr(part, "text", None):
-                                description_text = part.text
-                                break
-                    if description_text:
-                        break
-                if description_text:
-                    async with db_session() as session:
-                        session.add(ModelDescription(s3_key=key, description=description_text))
-            except Exception:
-                pass
-            return StreamingResponse(BytesIO(png_bytes), media_type="image/png")
-        return JSONResponse({"error": "no image from model"}, status_code=502)
-    except Exception as e:
-        logger.exception("model generate failed")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.post("/env/defaults")
-async def set_defaults(
-    s3_keys: list[str] = Form(...),
-    names: list[str] = Form(...),
-    x_user_id: str | None = Header(default=None, alias="X-User-Id"),
-):
-    try:
-        if not x_user_id:
-            return JSONResponse({"error": "missing user id"}, status_code=400)
-        if len(s3_keys) != len(names):
-            return JSONResponse({"error": "mismatched arrays"}, status_code=400)
-        if len(s3_keys) > 5:
-            return JSONResponse({"error": "max 5 defaults"}, status_code=400)
-        # Overwrite all defaults
-        async with db_session() as session:
-            await session.execute(text("DELETE FROM env_defaults_user WHERE user_id = :uid"), {"uid": x_user_id})
-            for key, name in zip(s3_keys, names):
-                session.add(EnvDefaultUser(user_id=x_user_id, s3_key=key, name=name.strip() or "Untitled"))
-        return {"ok": True}
-    except Exception as e:
-        logger.exception("Failed to set defaults")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.delete("/env/defaults")
-async def unset_default(s3_key: str):
-    """Remove a single default by s3_key, keeping others intact."""
-    try:
-        async with db_session() as session:
-            # Remove from any user's defaults
-            await session.execute(text("DELETE FROM env_defaults_user WHERE s3_key = :k"), {"k": s3_key})
-        return {"ok": True}
-    except Exception as e:
-        logger.exception("Failed to unset default")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.patch("/env/defaults")
-async def rename_default(s3_key: str = Form(...), name: str = Form(...), x_user_id: str | None = Header(default=None, alias="X-User-Id")):
-    """Rename a single default by s3_key."""
-    try:
-        if not x_user_id:
-            return JSONResponse({"error": "missing user id"}, status_code=400)
-        name = (name or "").strip() or "Untitled"
-        async with db_session() as session:
-            await session.execute(
-                text("UPDATE env_defaults_user SET name = :n WHERE user_id = :uid AND s3_key = :k"),
-                {"n": name, "uid": x_user_id, "k": s3_key},
-            )
-        return {"ok": True}
-    except Exception as e:
-        logger.exception("Failed to rename default")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-# --- Manage generated images (delete) ---
-
-@app.delete("/env/generated")
-async def delete_generated(s3_key: str):
-    """Delete a generated environment image from S3 and DB. If it's set as a default, remove that too."""
-    try:
-        # Delete S3 object
-        delete_objects([s3_key])
-        # Delete from DB (generations + possibly defaults)
-        async with db_session() as session:
-            await session.execute(text("DELETE FROM generations WHERE s3_key = :k"), {"k": s3_key})
-            await session.execute(text("DELETE FROM env_defaults_user WHERE s3_key = :k"), {"k": s3_key})
-        return {"ok": True}
-    except Exception as e:
-        logger.exception("Failed to delete generated image")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-# --- Pose sources upload and pose-only description generation ---
-
-@app.post("/pose/sources/upload")
-async def upload_pose_sources(files: list[UploadFile] = File(...)):
-    try:
-        stored = []
-        for f in files:
-            data = await f.read()
-            _, key = upload_pose_source_image(data, mime=f.content_type)
-            async with db_session() as session:
-                session.add(PoseSource(s3_key=key))
-            stored.append({"s3_key": key})
-        return {"ok": True, "count": len(stored), "items": stored}
-    except Exception as e:
-        logger.exception("Failed to upload pose sources")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.get("/pose/sources")
-async def list_pose_sources():
-    try:
-        async with db_session() as session:
-            stmt = select(PoseSource.s3_key).order_by(PoseSource.created_at.desc())
-            res = await session.execute(stmt)
-            items = [row[0] for row in res.all()]
-        return {"ok": True, "items": items}
-    except Exception as e:
-        logger.exception("Failed to list pose sources")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.delete("/pose/sources")
-async def delete_pose_sources():
-    try:
-        async with db_session() as session:
-            res = await session.execute(select(PoseSource.s3_key))
-            keys = [row[0] for row in res.all()]
-        delete_objects(keys)
-        async with db_session() as session:
-            await session.execute(text("DELETE FROM pose_sources"))
-            await session.execute(text("DELETE FROM pose_descriptions"))
-        return {"ok": True, "deleted": len(keys)}
-    except Exception as e:
-        logger.exception("Failed to delete pose sources")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.post("/pose/describe")
-async def generate_pose_descriptions():
-    """Generate pose-only textual descriptions for all uploaded pose sources that lack one."""
-    try:
-        # Fetch sources without a PoseDescription
-        async with db_session() as session:
-            src_rows = await session.execute(select(PoseSource.s3_key).order_by(PoseSource.created_at.desc()))
-            src_keys = [row[0] for row in src_rows.all()]
-            if not src_keys:
-                return {"ok": True, "generated": 0}
-            # Find which already have descriptions
-            have_rows = await session.execute(select(PoseDescription.s3_key))
-            have = {row[0] for row in have_rows.all()}
-        todo = [k for k in src_keys if k not in have]
-        if not todo:
-            return {"ok": True, "generated": 0}
-
-        outcome = enqueue_pose_descriptions(todo, wait=True)
-        results = outcome.get("results", [])
-        generated = sum(1 for item in results if isinstance(item, dict) and item.get("ok"))
-        payload = {
-            "ok": True,
-            "generated": generated,
-            "queued": len(todo),
-        }
-        if outcome.get("job_id"):
-            payload["job_id"] = outcome["job_id"]
-        if outcome.get("inline"):
-            payload["inline"] = True
-        if outcome.get("fallback"):
-            payload["fallback"] = True
-        return payload
-    except Exception as e:
-        logger.exception("Failed to generate pose descriptions")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.get("/pose/descriptions")
-async def list_pose_descriptions():
-    try:
-        async with db_session() as session:
-            stmt = select(PoseDescription.s3_key, PoseDescription.description, PoseDescription.created_at).order_by(PoseDescription.created_at.desc())
-            res = await session.execute(stmt)
-            items = [{"s3_key": r[0], "description": r[1], "created_at": r[2].isoformat()} for r in res.all()]
-        return {"ok": True, "items": items}
-    except Exception as e:
-        logger.exception("Failed to list pose descriptions")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-# --- Model generated listing and defaults management ---
-
-@app.get("/model/generated")
-async def list_model_generated(x_user_id: str | None = Header(default=None, alias="X-User-Id")):
-    try:
-        async with db_session() as session:
-            if x_user_id:
-                stmt = (
-                    select(Generation.s3_key, Generation.created_at, Generation.options_json)
-                    .where(Generation.pose.in_(["model-man", "model-woman"]))
-                    .where(text("(options_json->>'user_id') = :uid")).params(uid=x_user_id)
-                    .order_by(Generation.created_at.desc())
-                    .limit(200)
-                )
-            else:
-                stmt = (
-                    select(Generation.s3_key, Generation.created_at, Generation.options_json)
-                    .where(text("1=0"))
-                )
-            res = await session.execute(stmt)
-            rows = res.all()
-            items = []
-            for row in rows:
-                key = row[0]
-                created = row[1]
-                gender = (row[2] or {}).get("gender")
-                # fetch description if present
-                desc_text = None
-                try:
-                    dres = await session.execute(text("SELECT description FROM model_descriptions WHERE s3_key = :k"), {"k": key})
-                    drow = dres.first()
-                    if drow:
-                        desc_text = drow[0]
-                except Exception:
-                    pass
-                try:
-                    url = generate_presigned_get_url(key)
-                except Exception:
-                    url = None
-                items.append({"s3_key": key, "created_at": created.isoformat(), "gender": gender, "url": url, "description": desc_text})
-        return {"ok": True, "items": items}
-    except Exception as e:
-        logger.exception("Failed to list model generated images")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-# --- Listings (group garment source, settings, generated images, description) ---
-
-@app.post("/listing")
-async def create_listing(
-    image: UploadFile = File(...),
-    gender: str = Form("woman"),
-    environment: str = Form("studio"),
-    poses: list[str] = Form(None),
-    extra: str = Form(""),
-    env_default_s3_key: str | None = Form(None),
-    model_default_s3_key: str | None = Form(None),
-    use_model_image: str | None = Form(None),  # "true" | "false"
-    prompt_override: str | None = Form(None),
-    title: str | None = Form(None),
-    garment_type_override: str | None = Form(None),
-    x_user_id: str | None = Header(default=None, alias="X-User-Id"),
-):
-    try:
-        if not x_user_id:
-            return JSONResponse({"error": "missing user id"}, status_code=400)
-        if not image or not image.filename:
-            return JSONResponse({"error": "image file required"}, status_code=400)
-        raw_bytes = await image.read()
-        if len(raw_bytes) > 10 * 1024 * 1024:
-            return JSONResponse({"error": "image too large (max ~10MB)"}, status_code=413)
-        try:
-            src = Image.open(BytesIO(raw_bytes))
-        except Exception:
-            return JSONResponse({"error": "invalid or unsupported image format"}, status_code=400)
-        buf = BytesIO()
-        try:
-            src.convert("RGBA").save(buf, format="PNG")
-        finally:
-            try:
-                src.close()
-            except Exception:
-                pass
-        buf.seek(0)
-
-        # Persist original garment source
-        try:
-            _, src_key = upload_product_source_image(buf.getvalue(), mime="image/png")
-        except Exception as e:
-            return JSONResponse({"error": f"failed to persist source image: {e}"}, status_code=500)
-
-        settings = {
-            "gender": (gender or "").strip().lower(),
-            "environment": (environment or "").strip().lower(),
-            "poses": poses if isinstance(poses, list) else ([poses] if poses else []),
-            "extra": (extra or "").strip(),
-            "env_default_s3_key": env_default_s3_key,
-            "model_default_s3_key": model_default_s3_key,
-            "use_model_image": (str(use_model_image).lower() == "true") if use_model_image is not None else None,
-            "prompt_override": (prompt_override or "").strip() or None,
-            "title": (title or "").strip() or None,
-            "garment_type_override": (garment_type_override or None),
-        }
-
-        lid = uuid.uuid4().hex
-        async with db_session() as session:
-            session.add(
-                Listing(
-                    id=lid,
-                    user_id=x_user_id,
-                    source_s3_key=src_key,
-                    settings_json=settings,
-                    description_text=None,
-                    cover_s3_key=None,
-                )
-            )
-
-        try:
-            src_url = generate_presigned_get_url(src_key)
-        except Exception:
-            src_url = None
-
-        return {
-            "ok": True,
-            "id": lid,
-            "source_s3_key": src_key,
-            "source_url": src_url,
-            "settings": settings,
-        }
-    except Exception as e:
-        logger.exception("failed to create listing")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.get("/listings")
-async def list_listings(x_user_id: str | None = Header(default=None, alias="X-User-Id")):
-    try:
-        if not x_user_id:
-            return JSONResponse({"error": "missing user id"}, status_code=400)
-        async with db_session() as session:
-            # Fetch recent listings for user
-            lres = await session.execute(
-                text(
-                    "SELECT id, created_at, cover_s3_key, settings_json FROM listings "
-                    "WHERE user_id = :uid ORDER BY created_at DESC LIMIT 200"
-                ),
-                {"uid": x_user_id},
-            )
-            rows = lres.all()
-            # Get counts per listing in a second query
-            ids = [r[0] for r in rows]
-            counts: dict[str, int] = {}
-            if ids:
-                cres = await session.execute(
-                    text(
-                        "SELECT listing_id, COUNT(*) FROM listing_images "
-                        "WHERE listing_id = ANY(:ids) GROUP BY listing_id"
-                    ),
-                    {"ids": ids},
-                )
-                for lid, cnt in cres.all():
-                    counts[str(lid)] = int(cnt)
-        items = []
-        for lid, created_at, cover_key, settings_json in rows:
-            try:
-                cover_url = generate_presigned_get_url(cover_key) if cover_key else None
-            except Exception:
-                cover_url = None
-            items.append(
-                {
-                    "id": lid,
-                    "created_at": created_at.isoformat(),
-                    "cover_s3_key": cover_key,
-                    "cover_url": cover_url,
-                    "images_count": counts.get(str(lid), 0),
-                    "settings": settings_json or {},
-                }
-            )
-        return {"ok": True, "items": items}
-    except Exception as e:
-        logger.exception("failed to list listings")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.get("/listing/{lid}")
-async def get_listing(lid: str, x_user_id: str | None = Header(default=None, alias="X-User-Id")):
-    try:
-        if not x_user_id:
-            return JSONResponse({"error": "missing user id"}, status_code=400)
-        async with db_session() as session:
-            lres = await session.execute(
-                text(
-                    "SELECT id, user_id, source_s3_key, settings_json, description_text, cover_s3_key, created_at "
-                    "FROM listings WHERE id = :id"
-                ),
-                {"id": lid},
-            )
-            lrow = lres.first()
-            if not lrow or lrow[1] != x_user_id:
-                return JSONResponse({"error": "not found"}, status_code=404)
-            # Fetch images
-            ires = await session.execute(
-                text(
-                    "SELECT s3_key, pose, prompt, created_at FROM listing_images "
-                    "WHERE listing_id = :id ORDER BY created_at DESC"
-                ),
-                {"id": lid},
-            )
-            irows = ires.all()
-        # Build response
-        try:
-            source_url = generate_presigned_get_url(lrow[2]) if lrow[2] else None
-        except Exception:
-            source_url = None
-        try:
-            cover_url = generate_presigned_get_url(lrow[5]) if lrow[5] else None
-        except Exception:
-            cover_url = None
-        images = []
-        for s3_key, pose, prompt, created_at in irows:
-            try:
-                url = generate_presigned_get_url(s3_key)
-            except Exception:
-                url = None
-            images.append(
-                {
-                    "s3_key": s3_key,
-                    "pose": pose,
-                    "prompt": prompt,
-                    "created_at": created_at.isoformat(),
-                    "url": url,
-                }
-            )
-        return {
-            "ok": True,
-            "id": lrow[0],
-            "created_at": lrow[6].isoformat(),
-            "source_s3_key": lrow[2],
-            "source_url": source_url,
-            "settings": lrow[3] or {},
-            "description_text": lrow[4],
-            "cover_s3_key": lrow[5],
-            "cover_url": cover_url,
-            "images": images,
-        }
-    except Exception as e:
-        logger.exception("failed to get listing")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.patch("/listing/{lid}/cover")
-async def set_listing_cover(lid: str, s3_key: str = Form(...), x_user_id: str | None = Header(default=None, alias="X-User-Id")):
-    try:
-        if not x_user_id:
-            return JSONResponse({"error": "missing user id"}, status_code=400)
-        # Verify listing ownership and that the image belongs to this listing
-        async with db_session() as session:
-            lres = await session.execute(text("SELECT user_id FROM listings WHERE id = :id"), {"id": lid})
-            row = lres.first()
-            if not row or row[0] != x_user_id:
-                return JSONResponse({"error": "not found"}, status_code=404)
-            ires = await session.execute(
-                text("SELECT 1 FROM listing_images WHERE listing_id = :id AND s3_key = :k LIMIT 1"),
-                {"id": lid, "k": s3_key},
-            )
-            if not ires.first():
-                return JSONResponse({"error": "image not part of listing"}, status_code=400)
-            await session.execute(text("UPDATE listings SET cover_s3_key = :k WHERE id = :id"), {"k": s3_key, "id": lid})
-        return {"ok": True}
-    except Exception as e:
-        logger.exception("failed to set listing cover")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.post("/edit/json")
-async def edit_json(
-    image: UploadFile | None = File(None),
-    gender: str = Form("woman"),
-    environment: str = Form("studio"),
-    poses: list[str] = Form(None),
-    extra: str = Form(""),
-    env_default_s3_key: str | None = Form(None),
-    model_default_s3_key: str | None = Form(None),
-    model_description_text: str | None = Form(None),
-    prompt_override: str | None = Form(None),
-    listing_id: str | None = Form(None),
-    garment_type_override: str | None = Form(None),
-    x_user_id: str | None = Header(default=None, alias="X-User-Id"),
-):
-    """Same as /edit but returns JSON and optionally attaches to a listing."""
-    try:
-        # Load source image bytes either from uploaded file or from listing source
-        src_png: bytes | None = None
-        if image and image.filename:
-            raw_bytes = await image.read()
-            if len(raw_bytes) > 20 * 1024 * 1024:
-                return JSONResponse({"error": "image too large (max ~20MB)"}, status_code=413)
-            try:
-                src_png = _normalize_to_png_limited(raw_bytes, max_px=2048)
-            except Exception:
-                return JSONResponse({"error": "invalid or unsupported image format"}, status_code=400)
-        elif listing_id and x_user_id:
-            # Fetch listing and ensure ownership, then load its stored source image from S3
-            async with db_session() as session:
-                owns = await session.execute(
-                    text("SELECT user_id, source_s3_key FROM listings WHERE id = :id"),
-                    {"id": listing_id},
-                )
-                row = owns.first()
-            if not row or row[0] != x_user_id:
-                return JSONResponse({"error": "not found"}, status_code=404)
-            try:
-                src_bytes, src_mime = get_object_bytes(row[1])
-                # Normalize to PNG and downscale for consistency and stability
-                src_png = _normalize_to_png_limited(src_bytes, max_px=2048)
-            except Exception as e:
-                return JSONResponse({"error": f"failed to load source image from listing: {e}"}, status_code=500)
-        else:
-            return JSONResponse({"error": "image file or listing_id required"}, status_code=400)
-
-        # Normalize some fields
-        gender = _normalize_choice(gender, ["woman", "man"], "woman")
-        environment = _normalize_choice(environment, ["studio", "street", "bed", "beach", "indoor"], "studio")
-        if not poses:
-            poses = []
-        if not isinstance(poses, list):
-            poses = [poses]
-        # Keep first pose string for metadata; do not hard-normalize names here
-        pose_str = (poses[0] if poses else "") or ""
-        extra = (extra or "").strip()
-        if len(extra) > 200:
-            extra = extra[:200]
-
-        # Determine garment type
-        garment_type = await _classify_garment_type(src_png, garment_type_override)
-
-        # Build parts and call model
-        use_env_image = bool(env_default_s3_key)
-        use_person_image = bool(model_default_s3_key)
-        # Build prompt and parts in numbered image order: 1=garment, 2=person (opt), 3=environment (opt)
-        if prompt_override and prompt_override.strip():
-            prompt_text = prompt_override.strip()
-            prompt_variant = "override"
-        else:
-            prompt_text = classic_detailed(
-                gender=gender,
-                environment=environment,
-                pose=pose_str,
-                use_person_image=use_person_image,
-                use_env_image=use_env_image,
-                person_description=(model_description_text if (model_description_text and not use_person_image) else None),
-                garment_type=garment_type,
-            )
-            prompt_variant = "detailed"
-        parts: list[types.Part] = [types.Part.from_text(text=prompt_text)]
-        if (not use_person_image) and model_description_text:
-            parts.append(types.Part.from_text(text=f"Person description: {model_description_text}"))
-        # Image 1: garment
-        parts.append(types.Part.from_bytes(data=src_png, mime_type="image/png"))
-        env_key_used: str | None = None
-        person_key_used: str | None = None
-        # Image 2: person (opt)
-        if model_default_s3_key:
-            try:
-                person_bytes, person_mime = get_object_bytes(model_default_s3_key)
-                parts.append(types.Part.from_bytes(data=person_bytes, mime_type=person_mime or "image/png"))
-                person_key_used = model_default_s3_key
-            except Exception:
-                person_key_used = None
-        # Image 3: environment (opt)
-        if env_default_s3_key:
-            try:
-                env_bytes, env_mime = get_object_bytes(env_default_s3_key)
-                parts.append(types.Part.from_bytes(data=env_bytes, mime_type=env_mime or "image/png"))
-                env_key_used = env_default_s3_key
-            except Exception:
-                env_key_used = None
-
-        resp = await _genai_generate_with_retries(parts, attempts=2)
-        png_bytes = _first_inline_image_bytes(resp)
-        if png_bytes:
-            # Upload and persist generation
-            _, key = upload_image(png_bytes, pose=pose_str or "pose")
-            async with db_session() as session:
-                session.add(
-                    Generation(
-                        s3_key=key,
-                        pose=pose_str or "pose",
-                        prompt=prompt_text,
-                        options_json={
-                            "gender": gender,
-                            "environment": environment,
-                            "poses": poses,
-                            "extra": extra,
-                            "env_default_s3_key": env_key_used,
-                            "model_default_s3_key": person_key_used,
-                            "model_description_text": (model_description_text if not person_key_used else None),
-                            "garment_type": garment_type,
-                            "garment_type_override": (garment_type_override if garment_type_override else None),
-                            "user_id": x_user_id,
-                            "prompt_variant": prompt_variant,
-                        },
-                        model=MODEL,
-                    )
-                )
-                # Attach to listing if provided and owned by user
-                if listing_id and x_user_id:
-                    owns = await session.execute(
-                        text("SELECT 1 FROM listings WHERE id = :id AND user_id = :uid"),
-                        {"id": listing_id, "uid": x_user_id},
-                    )
-                    if owns.first():
-                        session.add(
-                            ListingImage(
-                                listing_id=listing_id,
-                                s3_key=key,
-                                pose=pose_str or "pose",
-                                prompt=prompt_text,
-                            )
-                        )
-                        # Set cover if not set yet
-                        await session.execute(
-                            text("UPDATE listings SET cover_s3_key = COALESCE(cover_s3_key, :k) WHERE id = :id"),
-                            {"k": key, "id": listing_id},
-                        )
-                        # Update listing settings with garment type and origin
-                        try:
-                            lres = await session.execute(
-                                text("SELECT settings_json FROM listings WHERE id = :id"),
-                                {"id": listing_id},
-                            )
-                            lrow = lres.first()
-                            settings = (lrow[0] or {}) if lrow else {}
-                            origin = "user" if (garment_type_override and garment_type_override.strip()) else "model"
-                            settings.update({
-                                "garment_type": garment_type,
-                                "garment_type_origin": origin,
-                            })
-                            await session.execute(
-                                text("UPDATE listings SET settings_json = :j WHERE id = :id"),
-                                {"j": settings, "id": listing_id},
-                            )
-                        except Exception:
-                            pass
-            try:
-                url = generate_presigned_get_url(key)
-            except Exception:
-                url = None
-            return {
-                "ok": True,
-                "s3_key": key,
-                "url": url,
-                "pose": pose_str or "pose",
-                "prompt": prompt_text,
-                "listing_id": listing_id,
-            }
-        # Fallback to concise prompt if no image and no explicit override
-        if not (prompt_override and prompt_override.strip()):
-            try:
-                prompt_variant = "concise"
-                prompt_text = classic_concise(
-                    gender=gender,
-                    environment=environment,
-                    pose=pose_str,
-                    use_person_image=use_person_image,
-                    use_env_image=use_env_image,
-                    person_description=(model_description_text if (model_description_text and not use_person_image) else None),
-                    garment_type=garment_type,
-                )
-                parts[0] = types.Part.from_text(text=prompt_text)
-                resp2 = await _genai_generate_with_retries(parts, attempts=1)
-                png_bytes = _first_inline_image_bytes(resp2)
-                if png_bytes:
-                    _, key = upload_image(png_bytes, pose=pose_str or "pose")
-                    async with db_session() as session:
-                        session.add(
-                            Generation(
-                                s3_key=key,
-                                pose=pose_str or "pose",
-                                prompt=prompt_text,
-                                options_json={
-                                    "gender": gender,
-                                    "environment": environment,
-                                    "poses": poses,
-                                    "extra": extra,
-                                    "env_default_s3_key": env_key_used,
-                                    "model_default_s3_key": person_key_used,
-                                    "model_description_text": (model_description_text if not person_key_used else None),
-                                    "garment_type": garment_type,
-                                    "garment_type_override": (garment_type_override if garment_type_override else None),
-                                    "user_id": x_user_id,
-                                    "prompt_variant": prompt_variant,
-                                },
-                                model=MODEL,
-                            )
-                        )
-                        # Attach to listing if provided and owned by user
-                        if listing_id and x_user_id:
-                            owns = await session.execute(
-                                text("SELECT 1 FROM listings WHERE id = :id AND user_id = :uid"),
-                                {"id": listing_id, "uid": x_user_id},
-                            )
-                            if owns.first():
-                                session.add(
-                                    ListingImage(
-                                        listing_id=listing_id,
-                                        s3_key=key,
-                                        pose=pose_str or "pose",
-                                        prompt=prompt_text,
-                                    )
-                                )
-                                # Update listing settings with garment type and origin
-                                try:
-                                    lres = await session.execute(
-                                        text("SELECT settings_json FROM listings WHERE id = :id"),
-                                        {"id": listing_id},
-                                    )
-                                    lrow = lres.first()
-                                    settings = (lrow[0] or {}) if lrow else {}
-                                    origin = "user" if (garment_type_override and garment_type_override.strip()) else "model"
-                                    settings.update({
-                                        "garment_type": garment_type,
-                                        "garment_type_origin": origin,
-                                    })
-                                    await session.execute(
-                                        text("UPDATE listings SET settings_json = :j WHERE id = :id"),
-                                        {"j": settings, "id": listing_id},
-                                    )
-                                except Exception:
-                                    pass
-                                # Set cover if not set yet
-                                await session.execute(
-                                    text("UPDATE listings SET cover_s3_key = COALESCE(cover_s3_key, :k) WHERE id = :id"),
-                                    {"k": key, "id": listing_id},
-                                )
-                    try:
-                        url = generate_presigned_get_url(key)
-                    except Exception:
-                        url = None
-                    return {
-                        "ok": True,
-                        "s3_key": key,
-                        "url": url,
-                        "pose": pose_str or "pose",
-                        "prompt": prompt_text,
-                        "listing_id": listing_id,
-                    }
-            except Exception:
-                pass
-        return JSONResponse({"error": "no edited image from model"}, status_code=502)
-    except genai_errors.APIError as e:
-        logger.exception("GenAI API error on /edit/json")
-        return JSONResponse({"error": e.message, "code": e.code}, status_code=502)
-    except Exception as e:
-        logger.exception("Unhandled error on /edit/json")
-        return JSONResponse({"error": str(e)}, status_code=500)
-    except Exception as e:
-        logger.exception("Failed to list user history")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.post("/edit/sequential/json")
-async def edit_sequential_json(
-    image: UploadFile | None = File(None),
-    gender: str = Form("woman"),
-    environment: str = Form("studio"),
-    poses: list[str] = Form(None),
-    extra: str = Form(""),
-    env_default_s3_key: str | None = Form(None),
-    model_default_s3_key: str | None = Form(None),
-    model_description_text: str | None = Form(None),
-    prompt_override_step1: str | None = Form(None),
-    prompt_override_step2: str | None = Form(None),
-    listing_id: str | None = Form(None),
-    garment_type_override: str | None = Form(None),
-    x_user_id: str | None = Header(default=None, alias="X-User-Id"),
-):
-    """Two-pass generation: (1) put garment on person; (2) place into scene.
-
-    Mirrors /edit/json inputs for drop-in replacement. Persists only final image.
-    """
-    try:
-        # Load source (garment) image
-        src_png: bytes | None = None
-        if image and image.filename:
-            raw_bytes = await image.read()
-            if len(raw_bytes) > 20 * 1024 * 1024:
-                return JSONResponse({"error": "image too large (max ~20MB)"}, status_code=413)
-            try:
-                src_png = _normalize_to_png_limited(raw_bytes, max_px=2048)
-            except Exception:
-                return JSONResponse({"error": "invalid or unsupported image format"}, status_code=400)
-        elif listing_id and x_user_id:
-            async with db_session() as session:
-                owns = await session.execute(
-                    text("SELECT user_id, source_s3_key FROM listings WHERE id = :id"),
-                    {"id": listing_id},
-                )
-                row = owns.first()
-            if not row or row[0] != x_user_id:
-                return JSONResponse({"error": "not found"}, status_code=404)
-            try:
-                src_bytes, _ = get_object_bytes(row[1])
-                src_png = _normalize_to_png_limited(src_bytes, max_px=2048)
-            except Exception as e:
-                return JSONResponse({"error": f"failed to load source image from listing: {e}"}, status_code=500)
-        else:
-            return JSONResponse({"error": "image file or listing_id required"}, status_code=400)
-
-        # Normalize inputs
-        if not poses:
-            poses = []
-        if not isinstance(poses, list):
-            poses = [poses]
-        pose_str = (poses[0] if poses else "") or ""
-        extra = (extra or "").strip()
-        if len(extra) > 200:
-            extra = extra[:200]
-        use_env_image = bool(env_default_s3_key)
-        use_person_image = bool(model_default_s3_key)
-        # Classify garment type for metadata and potential conditioning (currently used in stored options)
-        garment_type = await _classify_garment_type(src_png, garment_type_override)
-
-        # Step 1 prompt
-        if prompt_override_step1 and prompt_override_step1.strip():
-            step1_prompt = prompt_override_step1.strip()
-            step1_variant = "override"
-        else:
-            step1_prompt = seq_step1_detailed(
-                use_person_image=use_person_image,
-                pose=pose_str,
-                person_description=(model_description_text if (model_description_text and not use_person_image) else None),
-                gender=_normalize_choice(gender, ["woman", "man"], "woman"),
-            )
-            step1_variant = "detailed"
-        parts1: list[types.Part] = [types.Part.from_text(text=step1_prompt)]
-        person_key_used: str | None = None
-        if model_default_s3_key:
-            try:
-                person_bytes, person_mime = get_object_bytes(model_default_s3_key)
-                parts1.append(types.Part.from_text(text="Person reference:"))
-                parts1.append(types.Part.from_bytes(data=person_bytes, mime_type=person_mime or "image/png"))
-                person_key_used = model_default_s3_key
-            except Exception:
-                person_key_used = None
-        elif model_description_text:
-            parts1.append(types.Part.from_text(text=f"Person description: {model_description_text}"))
-            person_key_used = None
-        # Garment last
-        parts1.append(types.Part.from_bytes(data=src_png, mime_type="image/png"))
-
-        # Call step 1
-        resp1 = await _genai_generate_with_retries(parts1, attempts=2)
-        step1_png: bytes | None = _first_inline_image_bytes(resp1)
-        if not step1_png:
-            if not (prompt_override_step1 and prompt_override_step1.strip()):
-                try:
-                    step1_variant = "concise"
-                    step1_prompt = seq_step1_concise(
-                        use_person_image=use_person_image,
-                        pose=pose_str,
-                        person_description=(model_description_text if (model_description_text and not use_person_image) else None),
-                        gender=_normalize_choice(gender, ["woman", "man"], "woman"),
-                    )
-                    parts1[0] = types.Part.from_text(text=step1_prompt)
-                    resp1b = await _genai_generate_with_retries(parts1, attempts=1)
-                    step1_png = _first_inline_image_bytes(resp1b)
-                except Exception:
-                    step1_png = None
-            if not step1_png:
-                return JSONResponse({"error": "no image from model (step1)"}, status_code=502)
-
-        # Step 2 prompt
-        if prompt_override_step2 and prompt_override_step2.strip():
-            step2_prompt = prompt_override_step2.strip()
-            step2_variant = "override"
-        else:
-            step2_prompt = seq_step2_detailed(
-                use_env_image=use_env_image,
-                environment=_normalize_choice(environment, ["studio", "street", "bed", "beach", "indoor"], "studio"),
-                pose=pose_str,
-            )
-            step2_variant = "detailed"
-        parts2: list[types.Part] = [types.Part.from_text(text=step2_prompt)]
-        env_key_used: str | None = None
-        person_key_used: str | None = None
-        if env_default_s3_key:
-            try:
-                env_bytes, env_mime = get_object_bytes(env_default_s3_key)
-                parts2.append(types.Part.from_text(text="Environment reference:"))
-                parts2.append(types.Part.from_bytes(data=env_bytes, mime_type=env_mime or "image/png"))
-                env_key_used = env_default_s3_key
-            except Exception:
-                env_key_used = None
-        # Person reference is the step-1 result
-        parts2.append(types.Part.from_text(text="Person reference:"))
-        parts2.append(types.Part.from_bytes(data=step1_png, mime_type="image/png"))
-
-        resp2 = await _genai_generate_with_retries(parts2, attempts=2)
-        final_png: bytes | None = _first_inline_image_bytes(resp2)
-        if final_png:
-            # Persist final
-            pose_final = (pose_str or "pose").strip() or "pose"
-            pose_final = f"{pose_final} (seq)"  # A/B: suffix so UI can distinguish without schema changes
-            bucket, key = upload_image(final_png, pose=pose_final)
-            async with db_session() as session:
-                session.add(
-                    Generation(
-                        s3_key=key,
-                        pose=pose_final,
-                        prompt=step2_prompt,
-                        options_json={
-                            "flow": "sequential",
-                            "prompt_step1": step1_prompt,
-                            "prompt_step2": step2_prompt,
-                            "gender": gender,
-                            "environment": environment,
-                            "poses": poses,
-                            "extra": extra,
-                            "env_default_s3_key": env_key_used,
-                            "model_default_s3_key": person_key_used,
-                            "model_description_text": (model_description_text if not person_key_used else None),
-                            "garment_type": garment_type,
-                            "garment_type_override": (garment_type_override if garment_type_override else None),
-                            "user_id": x_user_id,
-                        },
-                        model=MODEL,
-                    )
-                )
-                if listing_id and x_user_id:
-                    owns = await session.execute(
-                        text("SELECT 1 FROM listings WHERE id = :id AND user_id = :uid"),
-                        {"id": listing_id, "uid": x_user_id},
-                    )
-                    if owns.first():
-                        session.add(
-                            ListingImage(
-                                listing_id=listing_id,
-                                s3_key=key,
-                                pose=pose_final,
-                                prompt=step2_prompt,
-                            )
-                        )
-                        await session.execute(
-                            text("UPDATE listings SET cover_s3_key = COALESCE(cover_s3_key, :k) WHERE id = :id"),
-                            {"k": key, "id": listing_id},
-                        )
-                        # Update listing settings with garment type and origin
-                        try:
-                            lres = await session.execute(
-                                text("SELECT settings_json FROM listings WHERE id = :id"),
-                                {"id": listing_id},
-                            )
-                            lrow = lres.first()
-                            settings = (lrow[0] or {}) if lrow else {}
-                            origin = "user" if (garment_type_override and garment_type_override.strip()) else "model"
-                            settings.update({
-                                "garment_type": garment_type,
-                                "garment_type_origin": origin,
-                            })
-                            await session.execute(
-                                text("UPDATE listings SET settings_json = :j WHERE id = :id"),
-                                {"j": settings, "id": listing_id},
-                            )
-                        except Exception:
-                            pass
-            try:
-                url = generate_presigned_get_url(key)
-            except Exception:
-                url = None
-            return {"ok": True, "s3_key": key, "url": url, "pose": pose_final, "prompt": step2_prompt, "listing_id": listing_id}
-        if not (prompt_override_step2 and prompt_override_step2.strip()):
-            try:
-                step2_variant = "concise"
-                step2_prompt = seq_step2_concise(
-                    use_env_image=use_env_image,
-                    environment=_normalize_choice(environment, ["studio", "street", "bed", "beach", "indoor"], "studio"),
-                    pose=pose_str,
-                )
-                parts2[0] = types.Part.from_text(text=step2_prompt)
-                resp2b = await _genai_generate_with_retries(parts2, attempts=1)
-                final_png = _first_inline_image_bytes(resp2b)
-                if final_png:
-                    pose_final = (pose_str or "pose").strip() or "pose"
-                    pose_final = f"{pose_final} (seq)"
-                    bucket, key = upload_image(final_png, pose=pose_final)
-                    async with db_session() as session:
-                        session.add(
-                            Generation(
-                                s3_key=key,
-                                pose=pose_final,
-                                prompt=step2_prompt,
-                                options_json={
-                                    "flow": "sequential",
-                                    "prompt_step1": step1_prompt,
-                                    "prompt_step2": step2_prompt,
-                                    "prompt_step1_variant": step1_variant,
-                                    "prompt_step2_variant": step2_variant,
-                                    "gender": gender,
-                                    "environment": environment,
-                                    "poses": poses,
-                                    "extra": extra,
-                                    "env_default_s3_key": env_key_used,
-                                    "model_default_s3_key": person_key_used,
-                                    "model_description_text": (model_description_text if not person_key_used else None),
-                                    "user_id": x_user_id,
-                                },
-                                model=MODEL,
-                            )
-                        )
-                        if listing_id and x_user_id:
-                            owns = await session.execute(
-                                text("SELECT 1 FROM listings WHERE id = :id AND user_id = :uid"),
-                                {"id": listing_id, "uid": x_user_id},
-                            )
-                            if owns.first():
-                                session.add(
-                                    ListingImage(
-                                        listing_id=listing_id,
-                                        s3_key=key,
-                                        pose=pose_final,
-                                        prompt=step2_prompt,
-                                    )
-                                )
-                                await session.execute(
-                                    text("UPDATE listings SET cover_s3_key = COALESCE(cover_s3_key, :k) WHERE id = :id"),
-                                    {"k": key, "id": listing_id},
-                                )
-                    try:
-                        url = generate_presigned_get_url(key)
-                    except Exception:
-                        url = None
-                    return {"ok": True, "s3_key": key, "url": url, "pose": pose_final, "prompt": step2_prompt, "listing_id": listing_id}
-            except Exception:
-                pass
-        return JSONResponse({"error": "no image from model (step2)"}, status_code=502)
-    except genai_errors.APIError as e:
-        logger.exception("GenAI API error on /edit/sequential/json")
-        return JSONResponse({"error": e.message, "code": e.code}, status_code=502)
-    except Exception as e:
-        logger.exception("Unhandled error on /edit/sequential/json")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-@app.get("/model/defaults")
-async def list_model_defaults():
-    try:
-        async with db_session() as session:
-            stmt = select(ModelDefault.gender, ModelDefault.s3_key, ModelDefault.name)
-            res = await session.execute(stmt)
-            rows = res.all()
-            # Preload descriptions for these keys in one query
-            keys = [row[1] for row in rows]
-            desc_map: dict[str, Optional[str]] = {}
-            if keys:
-                try:
-                    dstmt = select(ModelDescription.s3_key, ModelDescription.description).where(ModelDescription.s3_key.in_(keys))
-                    dres = await session.execute(dstmt)
-                    for k, d in dres.all():
-                        desc_map[k] = d
-                except Exception:
-                    # If description lookup fails, proceed without them
-                    desc_map = {}
-            items = []
-            for gender, key, name in rows:
-                try:
-                    url = generate_presigned_get_url(key)
-                except Exception:
-                    url = None
-                items.append({
-                    "gender": gender,
-                    "s3_key": key,
-                    "name": name,
-                    "url": url,
-                    "description": desc_map.get(key),
-                })
-        return {"ok": True, "items": items}
-    except Exception as e:
-        logger.exception("Failed to list model defaults")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.post("/model/defaults")
-async def set_model_default(gender: str = Form(...), s3_key: str = Form(...), name: str = Form("Default")):
-    try:
-        gender = _normalize_gender(gender)
-        async with db_session() as session:
-            # Upsert: keep only 1 per gender
-            await session.execute(text("DELETE FROM model_defaults WHERE gender = :g"), {"g": gender})
-            session.add(ModelDefault(gender=gender, s3_key=s3_key, name=(name or "").strip() or "Default"))
-        return {"ok": True}
-    except Exception as e:
-        logger.exception("Failed to set model default")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.patch("/model/defaults")
-async def rename_model_default(gender: str = Form(...), name: str = Form(...)):
-    try:
-        gender = _normalize_gender(gender)
-        name = (name or "").strip() or "Default"
-        async with db_session() as session:
-            await session.execute(text("UPDATE model_defaults SET name = :n WHERE gender = :g"), {"n": name, "g": gender})
-        return {"ok": True}
-    except Exception as e:
-        logger.exception("Failed to rename model default")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.delete("/model/defaults")
-async def unset_model_default(gender: str):
-    try:
-        gender = _normalize_gender(gender)
-        async with db_session() as session:
-            await session.execute(text("DELETE FROM model_defaults WHERE gender = :g"), {"g": gender})
-        return {"ok": True}
-    except Exception as e:
-        logger.exception("Failed to unset model default")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.delete("/model/generated")
-async def delete_model_generated(s3_key: str):
-    """Delete a generated model image from S3 and DB; clear from model defaults if set."""
-    try:
-        delete_objects([s3_key])
-        async with db_session() as session:
-            await session.execute(text("DELETE FROM generations WHERE s3_key = :k"), {"k": s3_key})
-            await session.execute(text("DELETE FROM model_defaults WHERE s3_key = :k"), {"k": s3_key})
-        return {"ok": True}
-    except Exception as e:
-        logger.exception("Failed to delete model generated image")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-# --- Model sources (admin UI helper) ---
-
-@app.post("/model/sources/upload")
-async def upload_model_sources(gender: str = Form(...), files: list[UploadFile] = File(...)):
-    """Upload one or more person source images for a given gender.
-
-    Stored in S3 under model_sources/<gender>/ and tracked in the DB.
-    """
-    try:
-        gender = _normalize_gender(gender)
-        stored: list[dict] = []
-        for f in files:
-            data = await f.read()
-            _, key = upload_model_source_image(data, gender=gender, mime=f.content_type)
-            async with db_session() as session:
-                session.add(ModelSource(gender=gender, s3_key=key))
-            stored.append({"gender": gender, "s3_key": key})
-        return {"ok": True, "count": len(stored), "items": stored}
-    except Exception as e:
-        logger.exception("Failed to upload model sources")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.get("/model/sources")
-async def list_model_sources(gender: str | None = None):
-    """List uploaded person source images. Optionally filter by gender."""
-    try:
-        async with db_session() as session:
-            stmt = select(ModelSource.gender, ModelSource.s3_key).order_by(ModelSource.created_at.desc())
-            if gender:
-                stmt = stmt.where(ModelSource.gender == _normalize_gender(gender))
-            res = await session.execute(stmt)
-            rows = res.all()
-            items = []
-            for g, key in rows:
-                try:
-                    url = generate_presigned_get_url(key)
-                except Exception:
-                    url = None
-                items.append({"gender": g, "s3_key": key, "url": url})
-        return {"ok": True, "items": items}
-    except Exception as e:
-        logger.exception("Failed to list model sources")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-# --- Product description generation (Vinted-style) ---
-
-@app.post("/describe")
-async def generate_product_description(
-    image: UploadFile = File(...),
-    gender: str = Form(""),
-    brand: str = Form(""),
-    model_name: str = Form(""),
-    size: str = Form(""),
-    condition: str = Form(""),
-    prompt_override: str | None = Form(None),
-    listing_id: str | None = Form(None),
-    x_user_id: str | None = Header(default=None, alias="X-User-Id"),
-):
-    """Generate a Vinted-style product description from an uploaded garment image and metadata."""
-    try:
-        if not image or not image.filename:
-            return JSONResponse({"error": "image file required"}, status_code=400)
-        # Normalize image
-        raw_bytes = await image.read()
-        if len(raw_bytes) > 10 * 1024 * 1024:
-            return JSONResponse({"error": "image too large (max ~10MB)"}, status_code=413)
-        try:
-            src = Image.open(BytesIO(raw_bytes))
-        except Exception:
-            return JSONResponse({"error": "invalid or unsupported image format"}, status_code=400)
-        buf = BytesIO()
-        try:
-            src.convert("RGBA").save(buf, format="PNG")
-        finally:
-            try:
-                src.close()
-            except Exception:
-                pass
-        buf.seek(0)
-
-        # Optional: persist original garment source for traceability
-        try:
-            _, src_key = upload_product_source_image(buf.getvalue(), mime="image/png")
-        except Exception:
-            src_key = None
-
-        # Build instruction (Gemini 2.5 Flash Image)
-        def norm(s: Optional[str]) -> str:
-            return (s or "").strip()
-
-        meta_lines = []
-        if norm(brand):
-            meta_lines.append(f"Brand: {norm(brand)}")
-        if norm(model_name):
-            meta_lines.append(f"Model: {norm(model_name)}")
-        if norm(size):
-            meta_lines.append(f"Size: {norm(size)}")
-        if norm(condition):
-            meta_lines.append(f"Condition: {norm(condition)}")
-        if norm(gender):
-            meta_lines.append(f"Gender: {norm(gender)}")
-
-        # Prefer explicit prompt_override from the client; else build a default instruction
-        if prompt_override and prompt_override.strip():
-            instruction = prompt_override.strip()
-        else:
-            instruction = (
-                "You are a helpful assistant that writes high-quality Vinted product listings from a product photo.\n"
-                "Output format EXACTLY as sections (plain text):\n"
-                "Title: <search-optimized; use brand, item, size, color, material, 1-2 key features; MAX 100 characters>\n\n"
-                "Description:\n"
-                "- Aim for 200–400 words (MIN 50 words; keep total under 3,000 characters).\n"
-                "- Use short paragraphs and hyphen bullets for measurements and unique features.\n"
-                "- Include brand, item type, size, color, material, fit, style keywords, unique features, and any visible flaws (be honest).\n"
-                "- Include measurements ONLY if clearly inferable; otherwise omit.\n\n"
-                "Condition: <one short line; use provided condition if present>\n"
-                "Extras: <one short line like 'Open to offers; bundle discounts' or leave empty>\n\n"
-                "Rules:\n"
-                "- Plain text only; no emojis; no markdown other than hyphen bullets; no price or shipping info.\n"
-                "- Keep PG-13.\n"
-            )
-            if meta_lines:
-                instruction += "\nKNOWN FIELDS (apply faithfully if present)\n" + "\n".join(meta_lines) + "\n"
-
-        parts: list[types.Part] = [
-            types.Part.from_text(text=instruction),
-            types.Part.from_bytes(data=buf.getvalue(), mime_type="image/png"),
-        ]
-        client = get_client()
-        resp = await asyncio.to_thread(
-            client.models.generate_content,
-            model=MODEL,
-            contents=types.Content(role="user", parts=parts),
-        )
-        description_text = None
-        for c in getattr(resp, "candidates", []) or []:
-            content = getattr(c, "content", None)
-            parts = getattr(content, "parts", None) if content is not None else None
-            if parts:
-                for p in parts:
-                    if getattr(p, "text", None):
-                        description_text = p.text
-                        break
-            if description_text:
-                break
-        if not description_text:
-            return JSONResponse({"error": "no description from model"}, status_code=502)
-
-        # Persist description
-        async with db_session() as session:
-            session.add(
-                ProductDescription(
-                    user_id=x_user_id,
-                    s3_key=src_key or "",
-                    gender=_normalize_gender(gender) if gender else None,
-                    brand=norm(brand) or None,
-                    model=norm(model_name) or None,
-                    size=norm(size) or None,
-                    condition=norm(condition) or None,
-                    description=description_text.strip(),
-                )
-            )
-            # If listing_id provided and owned by the user, store description on listing too
-            if listing_id and x_user_id:
-                owns = await session.execute(
-                    text("SELECT 1 FROM listings WHERE id = :id AND user_id = :uid"),
-                    {"id": listing_id, "uid": x_user_id},
-                )
-                if owns.first():
-                    await session.execute(
-                        text("UPDATE listings SET description_text = :d WHERE id = :id"),
-                        {"d": description_text.strip(), "id": listing_id},
-                    )
-
-        return {"ok": True, "description": description_text}
-    except Exception as e:
-        logger.exception("description generation failed")
-        return JSONResponse({"error": str(e)}, status_code=500)
-
-
-@app.post("/listing/{lid}/describe")
-async def describe_from_listing(
-    lid: str,
-    gender: str | None = Form(None),
-    brand: str | None = Form(None),
-    model_name: str | None = Form(None),
-    size: str | None = Form(None),
-    condition: str | None = Form(None),
-    x_user_id: str | None = Header(default=None, alias="X-User-Id"),
-):
-    """Generate a Vinted-style product description using the listing's stored source image.
-
-    Accepts optional fields (gender, brand, model_name, size, condition). Requires ownership.
-    """
-    try:
-        if not x_user_id:
-            return JSONResponse({"error": "missing user id"}, status_code=400)
-        # Fetch listing and verify ownership
-        async with db_session() as session:
-            res = await session.execute(
-                text("SELECT user_id, source_s3_key, settings_json FROM listings WHERE id = :id"),
-                {"id": lid},
-            )
-            row = res.first()
-        if not row or row[0] != x_user_id:
-            return JSONResponse({"error": "not found"}, status_code=404)
-        src_key = row[1]
-        settings = row[2] or {}
-
-        # Load source image from S3
-        try:
-            src_bytes, mime = get_object_bytes(src_key)
-        except Exception as e:
-            return JSONResponse({"error": f"failed to load source image: {e}"}, status_code=500)
-
-        def norm(s: Optional[str]) -> str:
-            return (s or "").strip()
-
-        # Merge provided fields with listing settings for gender default
-        gg = norm(gender) or norm(settings.get("gender"))
-        meta_lines = []
-        if norm(brand):
-            meta_lines.append(f"Brand: {norm(brand)}")
-        if norm(model_name):
-            meta_lines.append(f"Model: {norm(model_name)}")
-        if norm(size):
-            meta_lines.append(f"Size: {norm(size)}")
-        if norm(condition):
-            meta_lines.append(f"Condition: {norm(condition)}")
-        if gg:
-            meta_lines.append(f"Gender: {gg}")
-
-        instruction = (
-            "You are a helpful assistant that writes high-quality Vinted product listings from a product photo.\n"
-            "Output format EXACTLY as sections (plain text):\n"
-            "Title: <search-optimized; use brand, item, size, color, material, 1-2 key features; MAX 100 characters>\n\n"
-            "Description:\n"
-            "- Aim for 200–400 words (MIN 50 words; keep total under 3,000 characters).\n"
-            "- Use short paragraphs and hyphen bullets for measurements and unique features.\n"
-            "- Include brand, item type, size, color, material, fit, style keywords, unique features, and any visible flaws (be honest).\n"
-            "- Include measurements ONLY if clearly inferable; otherwise omit.\n\n"
-            "Condition: <one short line; use provided condition if present>\n"
-            "Extras: <one short line like 'Open to offers; bundle discounts' or leave empty>\n\n"
-            "Rules:\n"
-            "- Plain text only; no emojis; no markdown other than hyphen bullets; no price or shipping info.\n"
-            "- Keep PG-13.\n"
-        )
-        if meta_lines:
-            instruction += "\nKNOWN FIELDS (apply faithfully if present)\n" + "\n".join(meta_lines) + "\n"
-
-        parts: list[types.Part] = [
-            types.Part.from_text(text=instruction),
-            types.Part.from_bytes(data=src_bytes, mime_type=mime or "image/png"),
-        ]
-        client = get_client()
-        resp = await asyncio.to_thread(
-            client.models.generate_content,
-            model=MODEL,
-            contents=types.Content(role="user", parts=parts),
-        )
-        description_text = None
-        for c in getattr(resp, "candidates", []) or []:
-            content = getattr(c, "content", None)
-            p = getattr(content, "parts", None) if content is not None else None
-            if p:
-                for part in p:
-                    if getattr(part, "text", None):
-                        description_text = part.text
-                        break
-            if description_text:
-                break
-        if not description_text:
-            return JSONResponse({"error": "no description from model"}, status_code=502)
-
-        # Persist to ProductDescription and Listing
-        async with db_session() as session:
-            session.add(
-                ProductDescription(
-                    user_id=x_user_id,
-                    s3_key=src_key or "",
-                    gender=_normalize_gender(gg) if gg else None,
-                    brand=norm(brand) or None,
-                    model=norm(model_name) or None,
-                    size=norm(size) or None,
-                    condition=norm(condition) or None,
-                    description=description_text.strip(),
-                )
-            )
-            await session.execute(text("UPDATE listings SET description_text = :d WHERE id = :id"), {"d": description_text.strip(), "id": lid})
-
-        return {"ok": True, "description": description_text}
-    except Exception as e:
-        logger.exception("description from listing failed")
-        return JSONResponse({"error": str(e)}, status_code=500)
+async def on_shutdown() -> None:
+    await close_redis_client()
 
 
 if __name__ == "__main__":
