+++ backend/routes/__init__.py
@@ -1,0 +1,14 @@
+"""Route modules for the FastAPI backend."""
+from fastapi import APIRouter
+
+from . import admin, description, edit, environment, listing, model, pose
+
+
+router = APIRouter()
+router.include_router(admin.router)
+router.include_router(environment.router)
+router.include_router(model.router)
+router.include_router(pose.router)
+router.include_router(listing.router)
+router.include_router(edit.router)
+router.include_router(description.router)
